@@ -3,52 +3,60 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@material-ui/core": "^3.9.3",
-		"auth0-js": "^9.11.2",
-		"axios": "^0.18.0",
-		"body-parser": "^1.18.3",
-		"bootstrap": "^4.3.1",
-		"convert-excel-to-json": "^1.6.0",
+		"@material-ui/core": "^4.10.1",
+		"auth0-js": "^9.13.2",
+		"axios": "^0.19.2",
+		"body-parser": "^1.19.0",
+		"bootstrap": "^4.5.0",
+		"convert-excel-to-json": "^1.7.0",
 		"cors": "^2.8.5",
-		"emotion": "^10.0.6",
+		"date-fns": "^2.14.0",
+		"emotion": "^10.0.27",
+		"eslint-config-react-app": "^5.2.1",
+		"eslint-plugin-flowtype": "^5.1.3",
+		"eslint-plugin-import": "^2.21.2",
+		"eslint-plugin-react-hooks": "^4.0.4",
 		"excel-as-json": "^2.0.2",
-		"express-jwt": "^5.3.1",
+		"express-jwt": "^5.3.3",
 		"express-jwt-authz": "^2.3.1",
-		"express-validator": "^6.2.0",
+		"express-validator": "^6.5.0",
 		"jest-transform-stub": "^2.0.0",
-		"jwks-rsa": "^1.6.0",
+		"jwks-rsa": "^1.8.0",
 		"jwt-decode": "^2.2.0",
-		"moment": "^2.24.0",
-		"mysql": "^2.17.1",
-		"mysql2": "^1.7.0",
-		"node-fetch": "^2.3.0",
-		"node-sass": "^4.11.0",
-		"nodemon": "^1.19.1",
-		"react": "^16.7.0",
-		"react-data-table-component": "^3.4.1",
-		"react-datepicker": "^2.8.0",
-		"react-dom": "^16.7.0",
-		"react-popper": "^1.3.4",
-		"react-redux": "^6.0.0",
-		"react-router-dom": "^4.3.1",
-		"react-scripts": "2.1.3",
-		"react-slick": "^0.25.2",
-		"react-toastify": "^5.4.0",
-		"react-transition-group": "^4.3.0",
+		"moment": "^2.26.0",
+		"mysql": "^2.18.1",
+		"mysql2": "^2.1.0",
+		"node-fetch": "^2.6.0",
+		"node-sass": "^4.14.1",
+		"nodemon": "^2.0.4",
+		"react": "^16.13.1",
+		"react-confirm-alert": "^2.6.1",
+		"react-data-table-component": "^6.9.3",
+		"react-datepicker": "^3.0.0",
+		"react-dom": "^16.13.1",
+		"react-popper": "^2.2.3",
+		"react-redux": "^7.2.0",
+		"react-router-dom": "^5.2.0",
+		"react-scripts": "^3.4.1",
+		"react-slick": "^0.26.1",
+		"react-toastify": "^6.0.5",
+		"react-transition-group": "^4.4.1",
+		"react-uuid": "^1.0.2",
 		"reactjs-popup": "^1.5.0",
-		"reactn": "^0.1.9",
-		"reactstrap": "^8.0.1",
-		"redux": "^4.0.1",
+		"reactn": "^2.2.7",
+		"reactstrap": "^8.4.1",
+		"redux": "^4.0.5",
 		"redux-devtools-extension": "^2.13.8",
-		"redux-mock-store": "^1.5.3",
+		"redux-mock-store": "^1.5.4",
 		"redux-thunk": "^2.3.0",
-		"semantic-ui-react": "^0.88.1",
-		"sequelize": "^5.17.1",
-		"sequelize-cli": "^5.5.0",
+		"semantic-ui-react": "^0.88.2",
+		"sequelize": "^5.21.12",
+		"sequelize-cli": "^5.5.1",
 		"slick-carousel": "^1.8.1",
-		"styled-components": "^4.3.2",
-		"underscore": "^1.9.1",
-		"xls-to-json": "^0.5.1"
+		"styled-components": "^5.1.1",
+		"underscore": "^1.10.2",
+		"uuid": "^8.1.0",
+		"xls-to-json": "^0.5.4"
 	},
 	"scripts": {
 		"start": "webpack-dev-server --mode development --open --hot",
@@ -69,39 +79,36 @@
 		"not op_mini all"
 	],
 	"devDependencies": {
-		"@babel/cli": "^7.0.0",
-		"@babel/core": "^7.0.0",
-		"@babel/plugin-proposal-class-properties": "^7.0.0",
-		"@babel/preset-env": "^7.0.0",
-		"@babel/preset-react": "^7.0.0",
-		"babel-eslint": "10.0.1",
-		"babel-loader": "^8.0.2",
+		"@babel/cli": "^7.10.1",
+		"@babel/core": "^7.10.2",
+		"@babel/plugin-proposal-class-properties": "^7.10.1",
+		"@babel/preset-env": "^7.10.2",
+		"@babel/preset-react": "^7.10.1",
+		"babel-eslint": "^10.1.0",
+		"babel-loader": "^8.1.0",
 		"babel-plugin-transform-class-properties": "^6.24.1",
-		"concurrently": "^4.1.2",
-		"css-loader": "^1.0.0",
-		"enzyme": "^3.8.0",
-		"enzyme-adapter-react-16": "^1.9.1",
-		"eslint": "^5.12.1",
-		"eslint-config-airbnb": "17.1.0",
-		"eslint-config-airbnb-base": "13.1.0",
-		"eslint-config-prettier": "^3.1.0",
-		"eslint-import-resolver-webpack": "0.10.1",
-		"eslint-plugin-import": "2.14.0",
-		"eslint-plugin-jsx-a11y": "6.1.2",
-		"eslint-plugin-prettier": "^3.0.0",
-		"eslint-plugin-react": "7.11.1",
-		"file-loader": "^2.0.0",
-		"html-loader": "^0.5.5",
-		"html-webpack-plugin": "^3.2.0",
-		"jest": "^24.1.0",
-		"prettier": "^1.18.2",
-		"react-bootstrap": "^1.0.0-beta.12",
-		"sass-loader": "^6.0.7",
-		"style-loader": "^0.23.0",
-		"url-loader": "^1.1.1",
-		"webpack": "^4.28.4",
-		"webpack-bundle-tracker": "^0.3.0",
-		"webpack-cli": "^3.2.1",
-		"webpack-dev-server": "^3.7.2"
+		"concurrently": "^5.2.0",
+		"css-loader": "^3.5.3",
+		"enzyme": "^3.11.0",
+		"enzyme-adapter-react-16": "^1.15.2",
+		"eslint": "^7.2.0",
+		"eslint-config-airbnb": "^18.1.0",
+		"eslint-config-prettier": "^6.11.0",
+		"eslint-plugin-jsx-a11y": "^6.2.3",
+		"eslint-plugin-prettier": "^3.1.3",
+		"eslint-plugin-react": "^7.20.0",
+		"file-loader": "^6.0.0",
+		"html-loader": "^1.1.0",
+		"html-webpack-plugin": "^4.3.0",
+		"jest": "^26.0.1",
+		"prettier": "^2.0.5",
+		"react-bootstrap": "^1.0.1",
+		"sass-loader": "^8.0.2",
+		"style-loader": "^1.2.1",
+		"url-loader": "^4.1.0",
+		"webpack": "^4.43.0",
+		"webpack-bundle-tracker": "^1.0.0-alpha.1",
+		"webpack-cli": "^3.3.11",
+		"webpack-dev-server": "^3.11.0"
 	}
 }
