@@ -8,7 +8,7 @@ import { isEmpty } from "lodash";
 
 const modalStyles = {
   display: "flex",
-  alignItems: "center"
+  alignItems: "center",
 };
 
 /**
@@ -26,14 +26,15 @@ class DialogBox extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      confirmButtonDisable: false
+      confirmButtonDisable: false,
     };
   }
+
   initialState = props => {
     const { data, edit, create } = props;
 
-    const values = create ? data : data.values,
-      id = data.id;
+    const values = create ? data : data.values;
+    const { id } = data;
     let confirmButtonDisable = false;
     if (create) {
       confirmButtonDisable = true;
@@ -44,21 +45,24 @@ class DialogBox extends Component {
         type: "choice",
         title: "Choose new choice",
         value: "",
-        editable: true
+        editable: true,
       });
     }
-    this.setState({ data: values, edit, id, confirmButtonDisable });
+    this.setState({
+      data: values, edit, id, confirmButtonDisable,
+    });
   };
+
   componentDidMount = () => {
     this.initialState(this.props);
   };
-  componentWillReceiveProps = props => {
+
+  UNSAFE_componentWillReceiveProps = props => {
     this.initialState(props);
   };
 
-  convertChoice = value => {
-    return String.fromCharCode(value + 65);
-  };
+  convertChoice = value => String.fromCharCode(value + 65);
+
   onConfirm = () => {
     const { edit, create, onConfirm } = this.props;
     const { data, id } = this.state;
@@ -72,6 +76,7 @@ class DialogBox extends Component {
       onConfirm();
     }
   };
+
   valueChange = (value, title, index = 0) => {
     const { data } = this.state;
     data.map(item => {
@@ -81,34 +86,35 @@ class DialogBox extends Component {
           : (item.value = value);
         if (isEmpty(value)) {
           item.type === "options"
-            ? (item["error"] = true)
-            : (item["error"] = true);
+            ? (item.error = true)
+            : (item.error = true);
         } else {
-          item["error"] = false;
+          item.error = false;
         }
       }
     });
     this.setState({ data, confirmButtonDisable: isEmpty(value) });
   };
+
   render() {
     const {
-        bsSize,
-        cancelButtonType,
-        cancelButtonValue,
-        confirmButtonValue,
-        isConfirmation,
-        messageDescription,
-        messageNote,
-        messageTitle,
-        onCancel,
-        showMessage,
-        title,
-        messageBox,
-        create,
-        onDelete,
-        removeMessage
-      } = this.props,
-      { data, edit, confirmButtonDisable } = this.state;
+      bsSize,
+      cancelButtonType,
+      cancelButtonValue,
+      confirmButtonValue,
+      isConfirmation,
+      messageDescription,
+      messageNote,
+      messageTitle,
+      onCancel,
+      showMessage,
+      title,
+      messageBox,
+      create,
+      onDelete,
+      removeMessage,
+    } = this.props;
+    const { data, edit, confirmButtonDisable } = this.state;
     return (
       <div data-test="component-message-dialog">
         <Modal
@@ -119,7 +125,13 @@ class DialogBox extends Component {
         >
           <Modal.Header className="dialog-header">
             <div>{title}</div>
-            <div onClick={onCancel}>X</div>
+            <div
+              role="button"
+              tabIndex={0}
+              yonClick={onCancel}
+            >
+              X
+            </div>
           </Modal.Header>
           <Modal.Body className="dialog-body">
             {messageBox ? (
@@ -134,100 +146,100 @@ class DialogBox extends Component {
               </div>
             ) : (
               <div>
-                {!isEmpty(data) &&
-                  data.map((object, index) => {
+                {!isEmpty(data)
+                  && data.map((object, index) => {
                     switch (object.type) {
                       case "options":
                         return (
                           <div
-                            key={index}
+                            key={`option${index.toString()}`}
                             className="dialog-content dialog-options"
                           >
                             <div>
                               <div>Choice/Option</div>
                             </div>
                             <div>
-                              {object.value.map((option, option_index) => {
-                                return (
-                                  <div key={option_index}>
-                                    {this.convertChoice(option_index)} :
-                                    {create || edit ? (
-                                      <input
-                                        type="text"
-                                        name=""
-                                        value={option}
-                                        onChange={e =>
-                                          this.valueChange(
-                                            e.target.value,
-                                            object.title,
-                                            option_index
-                                          )
-                                        }
-                                      />
-                                    ) : (
-                                      <div>{option}</div>
-                                    )}
-                                  </div>
-                                );
-                              })}
+                              {object.value.map((option, option_index) => (
+                                <div key={option}>
+                                  {this.convertChoice(option_index)}
+                                  {' '}
+                                  :
+                                  {create || edit ? (
+                                    <input
+                                      type="text"
+                                      name=""
+                                      value={option}
+                                      onChange={e => this.valueChange(
+                                        e.target.value,
+                                        object.title,
+                                        option_index,
+                                      )}
+                                    />
+                                  ) : (
+                                    <div>{option}</div>
+                                  )}
+                                </div>
+                              ))}
                             </div>
                           </div>
                         );
                       case "choice":
                         return (
-                          <div key={index}>
+                          <div key={`choice${index.toString()}`}>
                             <div className="dialog-content">
-                              {object.title} :
+                              {object.title}
+                              {' '}
+                              :
                               <div className="choices">
                                 <div
+                                  role="button"
+                                  tabIndex={0}
                                   className={
                                     object.value === "A"
                                       ? "selected"
                                       : "unselected"
                                   }
-                                  onClick={() =>
-                                    object.editable &&
-                                    this.valueChange("A", object.title)
-                                  }
+                                  onClick={() => object.editable
+                                    && this.valueChange("A", object.title)}
                                 >
                                   A
                                 </div>
                                 <div
+                                  role="button"
+                                  tabIndex={0}
                                   className={
                                     object.value === "B"
                                       ? "selected"
                                       : "unselected"
                                   }
-                                  onClick={() =>
-                                    object.editable &&
-                                    this.valueChange("B", object.title)
-                                  }
+                                  onClick={() => object.editable
+                                    && this.valueChange("B", object.title)}
                                 >
                                   B
                                 </div>
                                 <div
+                                  role="button"
+                                  tabIndex={0}
                                   className={
                                     object.value === "C"
                                       ? "selected"
                                       : "unselected"
                                   }
-                                  onClick={() =>
-                                    object.editable &&
-                                    this.valueChange("C", object.title)
-                                  }
+                                  onClick={() => object.editable
+                                    && this.valueChange("C", object.title)}
                                 >
                                   C
                                 </div>
                                 <div
+                                  role="button"
+                                  tabIndex={0}
                                   className={
                                     object.value === "D"
                                       ? "selected"
                                       : "unselected"
                                   }
-                                  onClick={() =>
-                                    object.editable &&
-                                    this.valueChange("D", object.title)
-                                  }
+                                  onClick={() => object.editable
+                                    && this.valueChange("D", object.title)}
                                 >
                                   D
                                 </div>
@@ -237,35 +249,36 @@ class DialogBox extends Component {
                         );
                       case "text":
                         return (edit || create) && object.editable ? (
-                          <div key={index} className="dialog-content">
-                            {object.title} :
+                          <div key={`text${index.toString()}`} className="dialog-content">
+                            {object.title}
+                            {' '}
+                            :
                             {object.multiline ? (
                               <textarea
                                 name={object.title}
                                 className="dialog-multiline-text"
                                 value={object.value}
-                                onChange={e =>
-                                  this.valueChange(e.target.value, object.title)
-                                }
+                                onChange={e => this.valueChange(e.target.value, object.title)}
                               />
                             ) : (
                               <input
                                 type="text"
                                 name="object.title"
                                 value={object.value}
-                                onChange={e =>
-                                  this.valueChange(e.target.value, object.title)
-                                }
+                                onChange={e => this.valueChange(e.target.value, object.title)}
                               />
                             )}
                           </div>
                         ) : (
-                          <div key={index} className="dialog-content">
-                            {object.title} : {object.value}
+                          <div key={`text${index.toString()}`} className="dialog-content">
+                            {object.title}
+                            {' '}
+                            :
+                            {object.value}
                           </div>
                         );
                       default:
-                        return <div></div>;
+                        return <div />;
                     }
                   })}
                 {removeMessage && <div>{removeMessage}</div>}
@@ -304,18 +317,24 @@ class DialogBox extends Component {
 
 DialogBox.propTypes = {
   bsSize: PropTypes.string,
-  data: PropTypes.object,
+  data: PropTypes.shape({}),
   cancelButtonType: PropTypes.string,
   cancelButtonValue: PropTypes.string,
-  confirmButtonValue: PropTypes.string.isRequired,
+  confirmButtonValue: PropTypes.string,
   isConfirmation: PropTypes.bool,
   messageDescription: PropTypes.string,
   messageHeader: PropTypes.string,
   messageNote: PropTypes.string,
   messageTitle: PropTypes.string,
-  onCancel: PropTypes.func,
-  onConfirm: PropTypes.func,
-  showMessage: PropTypes.bool
+  onCancel: PropTypes.func.isRequired,
+  onConfirm: PropTypes.func.isRequired,
+  showMessage: PropTypes.bool,
+  messageBox: PropTypes.bool,
+  create: PropTypes.bool,
+  removeMessage: PropTypes.bool,
+  edit: PropTypes.bool.isRequired,
+  onDelete: PropTypes.func,
+  title: PropTypes.string.isRequired,
 };
 
 DialogBox.defaultProps = {
@@ -329,7 +348,11 @@ DialogBox.defaultProps = {
   messageHeader: "",
   messageNote: "",
   messageTitle: "",
-  showMessage: false
+  showMessage: false,
+  messageBox: false,
+  removeMessage: false,
+  create: false,
+  onDelete: null,
 };
 
 export default DialogBox;
