@@ -1,7 +1,4 @@
 import React from "react";
-import NdiLogoUrl from "../../images/ndiLogo.png";
-import infoUrl from "../../images/info.png";
-import profileUrl from "../../images/profile.png";
 import { ModuleCard } from "../../components/ModuleCard";
 import "../../commonStyles.scss";
 import "./styles.scss";
@@ -11,21 +8,12 @@ import {
   fetchGameData,
   fetchScores,
   fetchAuthDetails,
-  clearAuthDetails
+  clearAuthDetails,
 } from "./actions";
 import PropTypes from "prop-types";
 import GameInfo from "../../components/GameInfo";
-import * as jwtDecode from "jwt-decode";
 import Auth from "../../Auth";
-import { bindActionCreators } from "redux";
-//import {connect} from 'react-redux';
-import { Link } from "react-router-dom";
-import profile from "../../components/ProfileInfo";
-import admin from "../../components/admin/admin";
-import UpdatePlayer from "../../components/Update/UpdateProfile";
 import { fetchScoreDetail } from "../../components/ProfileInfo/action";
-import { da } from "date-fns/locale";
-import ProfileHeader from "../../components/ProfileHeader/ProfileHeader";
 
 const auth0 = new Auth();
 
@@ -35,7 +23,7 @@ const authDetail = {
   player_email: "",
   player_username: "",
   player_picture: "",
-  player_gender: ""
+  player_gender: "",
 };
 
 const scoreDetail = {
@@ -47,7 +35,7 @@ const scoreDetail = {
   program: null,
   total: 0,
   program_rank: null,
-  total_rank: null
+  total_rank: null,
 };
 
 global.fetch = require("node-fetch");
@@ -55,32 +43,30 @@ global.fetch = require("node-fetch");
 class LandingPage extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { 
+    this.state = {
       open: false,
-      color: "blue"
+      color: "blue",
     };
 
     this.colorChange = this.colorChange.bind(this);
 
-    //this.handleLogIn=this.handleLogIn.bind(this);
+    // this.handleLogIn=this.handleLogIn.bind(this);
   }
 
-  //Fetch complete game data.
-  componentWillMount() {
-
-    var cohort = 'default';
-    if(this.props.match) {
-     console.log(this.props.match.params);
-     cohort = this.props.match.params.cohortName ? this.props.match.params.cohortName:'default';
+  // Fetch complete game data.
+  UNSAFE_componentWillMount() {
+    let cohort = 'default';
+    if (this.props.match) {
+      cohort = this.props.match.params.cohortName ? this.props.match.params.cohortName : 'default';
     }
     // fetch('./moduleData.json')
-    fetch(config.baseUrl + `/api/v2/game/${cohort}`, {
+    fetch(`${config.baseUrl}/api/v2/game/${cohort}`, {
       method: "get",
       headers: {
-        authorization: "Bearer " + auth0.getAccessToken(),
+        authorization: `Bearer ${auth0.getAccessToken()}`,
         "Content-Type": "Application/json",
-        Accept: "application/json"
-      }
+        Accept: "application/json",
+      },
     })
       .then(response => {
         if (response.status >= 200 && response.status < 300) {
@@ -90,12 +76,12 @@ class LandingPage extends React.Component {
             this.props.getScores(scores);
           });
         } else if (response.status === 404) {
-          console.log("Not Found");
+          console.log("Not Found"); // eslint-disable-line
         }
       })
-      .catch(err => console.log(err));
+      .catch(err => console.log(err)); // eslint-disable-line
 
-    //auth0.handleAuthentication();
+    // auth0.handleAuthentication();
     if (auth0.isAuthenticated() === true) {
       authDetail.player_given_name = auth0.getProfile().given_name;
       authDetail.player_family_name = auth0.getProfile().family_name;
@@ -104,73 +90,60 @@ class LandingPage extends React.Component {
       authDetail.player_email = auth0.getProfile().email;
       authDetail.player_picture = auth0.getProfile().picture;
       authDetail.player_gender = auth0.getProfile().gender;
-      console.log("auth details below: ----------------- ");
-      console.log(auth0.getProfile());
-      console.log(auth0.getProfile()["http://demGames.net/roles"]);
-      console.log(authDetail);
-      console.log(auth0.getProfile());
 
       this.props.setAuth(authDetail);
 
-      fetch(config.baseUrl + "/user/findOne/" + authDetail.player_email, {
+      fetch(`${config.baseUrl}/user/findOne/${authDetail.player_email}`, {
         method: "get",
         headers: {
-          authorization: "Bearer " + auth0.getAccessToken(),
+          authorization: `Bearer ${auth0.getAccessToken()}`,
           "Content-Type": "Application/json",
-          Accept: "application/json"
-        }
+          Accept: "application/json",
+        },
       })
         .then(res => res.json())
         .then(data => {
-          console.log("user data from api below --V");
-          console.log(data);
-
           if (!data.email) {
-            console.log("email not found --V");
+            console.log("email not found --V"); // eslint-disable-line
 
-            fetch(config.baseUrl + "/registerplayer", {
+            fetch(`${config.baseUrl}/registerplayer`, {
               method: "POST",
               headers: {
-                authorization: "Bearer " + auth0.getAccessToken(),
+                authorization: `Bearer ${auth0.getAccessToken()}`,
                 "Content-Type": "Application/json",
-                Accept: "application/json"
+                Accept: "application/json",
               },
               body: JSON.stringify({
                 firstName: authDetail.player_username,
                 email: authDetail.player_email,
-                userName: authDetail.player_username
-              })
+                userName: authDetail.player_username,
+              }),
             })
               .then(res => res.json())
-              .then(data => {
-                console.log("new player registered  ---V");
-                console.log(data);
-              })
+              .then(responseData => responseData)
               .catch(error => {
-                console.log(error);
+                console.log(error); // eslint-disable-line
               });
           }
         })
         .catch(error => {
-          console.log(error);
+          console.log(error); // eslint-disable-line
         });
     } else {
       this.props.clearAuth(authDetail);
     }
 
     if (auth0.isAuthenticated() === true) {
-      fetch(config.baseUrl + "/selectPlayerProfile", {
+      fetch(`${config.baseUrl}/selectPlayerProfile`, {
         method: "post",
         headers: {
           "Content-Type": "Application/json",
-          Accept: "application/json"
+          Accept: "application/json",
         },
-        body: JSON.stringify(auth0.getProfile())
+        body: JSON.stringify(auth0.getProfile()),
       })
         .then(res => res.json())
         .then(data => {
-          console.log(data);
-
           scoreDetail.current = data[0].current;
           scoreDetail.game_id = data[0].game_id;
           scoreDetail.play_id = data[0].id;
@@ -180,43 +153,41 @@ class LandingPage extends React.Component {
           scoreDetail.total_rank = data[0].total_rank;
           scoreDetail.program = data[0].program;
           scoreDetail.program_rank = data[0].program_rank;
-          console.log(scoreDetail);
           this.props.setScoreDetail(scoreDetail);
         })
-        .catch(error => console.log(error));
+        .catch(error => console.log(error)); // eslint-disable-line
     }
   }
-  //Fetch scores for each levels of each module.
+
+  // Fetch scores for each levels of each module.
   getScores = () => {
     const allScores = [];
     this.props.gameData.gameData.map(modules => {
       allScores.push(
-        modules.levels.map(level => {
-          return level.current_score;
-        })
+        modules.levels.map(level => level.current_score),
       );
     });
     return allScores;
   };
 
-  //Handle info icon click to open info dialog box.
+  // Handle info icon click to open info dialog box.
   handleClickOpen = () => {
     this.setState({ open: true });
   };
 
-  //Handle info dialog box close.
+  // Handle info dialog box close.
   handleClose = () => {
     this.setState({ open: false });
   };
 
-  //handle Login in action
+  // handle Login in action
   handleLogIn = () => {
     if (!auth0.isAuthenticated()) {
       auth0.login();
     }
   };
 
-  //handle Logout in action
+  // handle Logout in action
   handleLogOut = () => {
     if (auth0.isAuthenticated()) {
       authDetail.player_given_name = "";
@@ -225,35 +196,27 @@ class LandingPage extends React.Component {
       authDetail.player_username = "";
       authDetail.player_picture = "";
       authDetail.player_gender = "";
-      console.log(authDetail);
       this.props.clearAuth(authDetail);
       auth0.logout();
     }
   };
 
-  // colorChange = (color) => {
-    
-  //   alert(color);
-    
-  // }
+ colorChange = (color) => {
+   this.setState({ color });
+ };
 
- colorChange = color => {
-    this.setState({color:color});
-  };
-
-  render() {
-    // console.log(auth0.getProfile()["http://demGames.net/roles"]&&auth0.getProfile()["http://demGames.net/roles"][0]==="admin");
-    const gameData = this.props.gameData.gameData;
-    const { open } = this.state;
-    return (
-      <div className="landing-page-wrapper">
-        <div className="landing-page-container">
-          <div className="game-title-container">
-            <p className="game-title">DemGames - Demo</p>
-          </div>
-          <div className="game-type-card-container">
-            {gameData.length > 0 &&
-              gameData.map((modules, key) => (
+ render() {
+   const { gameData } = this.props.gameData;
+   const { open } = this.state;
+   return (
+     <div className="landing-page-wrapper">
+       <div className="landing-page-container">
+         <div className="game-title-container">
+           <p className="game-title">DemGames - Demo</p>
+         </div>
+         <div className="game-type-card-container">
+           {gameData.length > 0
+              && gameData.map((modules) => (
                 <ModuleCard
                   key={modules.id}
                   moduleId={modules.id}
@@ -269,43 +232,61 @@ class LandingPage extends React.Component {
                   }
                 />
               ))}
-          </div>
-        </div>
-        {open && <GameInfo open={open} handleClose={this.handleClose} />}
-      </div>
-    );
-  }
+         </div>
+       </div>
+       {open && <GameInfo open={open} handleClose={this.handleClose} />}
+     </div>
+   );
+ }
 }
 
 const mapStateToProps = state => ({
   player_given_name: state.authDetail.authDetail.player_given_name,
   player_picture: state.authDetail.authDetail.player_picture,
-  gameData: state.gameData
+  gameData: state.gameData,
 });
 
-//Dispatch action to fetch game data and scores.
-const mapDispatchToProps = dispatch => {
-  console.log(scoreDetail);
-  return {
-    getGameData: gameData => dispatch(fetchGameData(gameData)),
-    getScores: scores => dispatch(fetchScores(scores)),
-    setAuth: authDetail => dispatch(fetchAuthDetails(authDetail)),
-    clearAuth: authDetail => dispatch(clearAuthDetails(authDetail)),
-    setScoreDetail: scoreDetail => dispatch(fetchScoreDetail(scoreDetail))
-  };
-};
+// Dispatch action to fetch game data and scores.
+const mapDispatchToProps = dispatch => ({
+  getGameData: gameData => dispatch(fetchGameData(gameData)),
+  getScores: scores => dispatch(fetchScores(scores)),
+  setAuth: () => dispatch(fetchAuthDetails(authDetail)),
+  clearAuth: () => dispatch(clearAuthDetails(authDetail)),
+  setScoreDetail: () => dispatch(fetchScoreDetail(scoreDetail)),
+});
 
 LandingPage.propTypes = {
   getGameData: PropTypes.func,
   getScores: PropTypes.func,
-  gameData: PropTypes.object,
-  authDetail: PropTypes.object,
+  gameData: PropTypes.shape({
+    gameData: PropTypes.arrayOf(PropTypes.shape({})),
+  }),
+  authDetail: PropTypes.shape({}),
   setAuth: PropTypes.func,
   clearAuth: PropTypes.func,
-  scoreDetail: PropTypes.object
+  setScoreDetail: PropTypes.func.isRequired,
+  scoreDetail: PropTypes.shape({}),
+  player_email: PropTypes.string,
+  match: PropTypes.shape({
+    params: PropTypes.shape({
+      cohortName: PropTypes.string,
+    }),
+  }),
+};
+
+LandingPage.defaultProps = {
+  getGameData: null,
+  getScores: null,
+  gameData: null,
+  authDetail: null,
+  setAuth: null,
+  clearAuth: null,
+  scoreDetail: null,
+  player_email: null,
+  match: null,
 };
 
 export default connect(
   mapStateToProps,
-  mapDispatchToProps
+  mapDispatchToProps,
 )(LandingPage);
