@@ -1,20 +1,14 @@
-import React, { Fragment } from "react";
-import LandingPage from "./pages/LandingPage/LandingPage";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import Alert from "./components/layout/alert";
-import "bootstrap/dist/css/bootstrap.css";
+import React from "react";
 // Redux
 import { Provider } from "react-redux";
+import LandingPage from "./pages/LandingPage/LandingPage";
+import "bootstrap/dist/css/bootstrap.css";
 import store from "./store";
 
-const App = () => {
-  return (
-    <Provider store={store}>
-      <Fragment>
-        <LandingPage />
-      </Fragment>
-    </Provider>
-  );
-};
+const App = () => (
+  <Provider store={store}>
+    <LandingPage />
+  </Provider>
+);
 
 export default App;
