@@ -1,16 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
 import { Provider } from 'react-redux';
+import App from './App';
 import store from './store';
 
 it('renders without crashing', () => {
-	const div = document.createElement('div');
-	ReactDOM.render(
-		<Provider store={store}>
-			<App />
-		</Provider>,
-		div
-	);
-	ReactDOM.unmountComponentAtNode(div);
+  const div = document.createElement('div');
+  ReactDOM.render(
+    <Provider store={store}>
+      <App />
+    </Provider>,
+    div,
+  );
+  ReactDOM.unmountComponentAtNode(div);
 });
