@@ -1,13 +1,11 @@
 import React, { useState } from "react";
-//import '../App.css';
 import { connect } from "react-redux";
-import PropTypes from "prop-types";
 
 const AddGame = () => {
   const initialState = {
     caption: "",
     gamedescription: "",
-    gametype: "multiplechoice"
+    gametype: "multiplechoice",
   };
 
   const [formData, setFormData] = useState(initialState);
@@ -42,14 +40,13 @@ const AddGame = () => {
   };
 
   const handleSubmit = event => {
-    console.log(formData);
     event.preventDefault();
   };
 
   return (
     <div className="container App">
       <div className="row">
-        <div className="col-md-2"></div>
+        <div className="col-md-2" />
         <div className="col-md-8">
           <form className="mt-5" onSubmit={e => handleSubmit(e)}>
             <input
@@ -59,7 +56,8 @@ const AddGame = () => {
               name="caption"
               value={caption}
               onChange={e => handleChange(e)}
-            />{" "}
+            />
+            {" "}
             <br />
             <textarea
               className="form-control mb-1"
@@ -68,28 +66,33 @@ const AddGame = () => {
               name="gamedescription"
               value={gamedescription}
               onChange={e => handleChange(e)}
-            ></textarea>
+            />
             <div className="row">
               <div className="form-group col-md-12">
-                <label className="form-label">Game type</label>
-                <select
-                  className="form-control custom-select custom-select-sm"
-                  name="gametype"
-                  value={gametype}
-                  onChange={e => handleChange(e)}
-                >
-                  <option value="multiplechoice">Multiple choice</option>
-                  <option value="matching">Matching</option>
-                  <option value="truefalse">True or False</option>
-                  <option value="fillin">Fill in</option>
-                </select>
+                <label htmlFor="gametype-select" className="form-label">
+                  <span>Game type </span>
+                  <select
+                    id="gametype-select"
+                    className="form-control custom-select custom-select-sm"
+                    name="gametype"
+                    value={gametype}
+                    onChange={e => handleChange(e)}
+                  >
+                    <option value="multiplechoice">Multiple choice</option>
+                    <option value="matching">Matching</option>
+                    <option value="truefalse">True or False</option>
+                    <option value="fillin">Fill in</option>
+                  </select>
+                </label>
               </div>
             </div>
             <div className="text-center mt-12">
               <button className="btn btn-info" type="submit">
                 Save
-              </button>{" "}
-              |{" "}
+              </button>
+              {" "}
+              |
+              {" "}
               <button
                 className="btn btn-warning"
                 type="button"
@@ -100,7 +103,7 @@ const AddGame = () => {
             </div>
           </form>
         </div>
-        <div className="col-md-2"></div>
+        <div className="col-md-2" />
       </div>
     </div>
   );
@@ -108,5 +111,5 @@ const AddGame = () => {
 
 export default connect(
   null,
-  {}
+  {},
 )(AddGame);
