@@ -1,9 +1,8 @@
 import React from "react";
+import PropTypes from "prop-types";
 import Slider from "react-slick";
 import "./styles.scss";
 import SliderArrow from "./SliderArrow";
-// import "slick-carousel/slick/slick.css";
-// import "slick-carousel/slick/slick-theme.css";
 
 const settings = {
   dots: false,
@@ -14,28 +13,42 @@ const settings = {
   prevArrow: <SliderArrow to="prev" />,
   nextArrow: <SliderArrow to="next" />,
 };
-export const Gamebox = ({ games, activeGame, handleGameBoxClick, deleteGame }) => {
-  console.log(games, "games");
-  return (
-    <div className="list-games-wrapper">
-      <div className="list-games-container">
-        <Slider {...settings}>
-          {games.length > 0 ? (
-            games.map(({ caption, id }, index) => (
-              <div className={`gamebox-wrapper ${activeGame === id ? "active" : ""}`}>
+export const Gamebox = ({
+  games, activeGame, handleGameBoxClick, deleteGame,
+}) => (
+  <div className="list-games-wrapper">
+    <div className="list-games-container">
+      <Slider {...settings}>
+        {games.length > 0 ? (
+          games.map(({ caption, id }, index) => (
+            <div key={`${id}${index.toString()}`} className={`gamebox-wrapper ${activeGame === id ? "active" : ""}`}>
               <div
+                role="button"
+                tabIndex={0}
                 className={`gamebox ${activeGame === id ? "active" : ""}`}
                 onClick={() => handleGameBoxClick(index)}
               >
-                <span className="game-delete" onClick={(e)=>{e.stopPropagation(); deleteGame(id)}}>x</span>
+                <span role="button" tabIndex={0} className="game-delete" onClick={(e) => { e.stopPropagation(); deleteGame(id); }}>
+                  x
+                </span>
                 <div className="game-title">{caption}</div>
               </div>
               {activeGame === id && <hr />}
-              </div>
-            ))
-          ) :null}
-        </Slider>
-      </div>
+            </div>
+          ))
+        ) : null}
+      </Slider>
     </div>
-  );
+  </div>
+);
+
+Gamebox.propTypes = {
+  games: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
+  activeGame: PropTypes.number,
+  handleGameBoxClick: PropTypes.func.isRequired,
+  deleteGame: PropTypes.func.isRequired,
+};
+
+Gamebox.defaultProps = {
+  activeGame: null,
 };
