@@ -1,38 +1,34 @@
-import React from "react";
-import NdiLogoUrl from "../../images/ndiLogo.png";
+import React from 'react';
+import PropTypes from 'prop-types';
 import "./styles.scss";
-import profileUrl from "../../images/profile.png";
-import Icon from "@material-ui/core/Icon";
-import { Link } from "react-router-dom";
 
-export const Header = ({ headerTabs, toggleTab, activeTab, name, image }) => {
-  console.log(name, image, "header");
-  return (
-    <div className="header-wrapper">
-      <div className="header-container">
-        <div className="header-left">
-          <div className="header-buttons">
-            {headerTabs.map(data => (
-              <button
-                className={`header-button ${
-                  activeTab.toLowerCase() === data.toLowerCase() ? "active" : ""
-                }`}
-                onClick={() => toggleTab(data)}
-              >
-                {data}
-              </button>
-            ))}
-          </div>
+export const Header = ({
+  headerTabs, toggleTab, activeTab,
+}) => (
+  <div className="header-wrapper">
+    <div className="header-container">
+      <div className="header-left">
+        <div className="header-buttons">
+          {headerTabs.map(data => (
+            <button
+              key={data}
+              type="submit"
+              className={`header-button ${
+                activeTab.toLowerCase() === data.toLowerCase() ? "active" : ""
+              }`}
+              onClick={() => toggleTab(data)}
+            >
+              {data}
+            </button>
+          ))}
         </div>
-        {/* <div className="header-right">
-          <div className="header-search">
-            <div className="header-searchbox">
-              <Icon className="search-icon">search</Icon>
-              <input className="header-input" placeholder="search"></input>
-            </div>
-          </div>
-        </div> */}
       </div>
     </div>
-  );
+  </div>
+);
+
+Header.propTypes = {
+  headerTabs: PropTypes.arrayOf(PropTypes.string).isRequired,
+  toggleTab: PropTypes.func.isRequired,
+  activeTab: PropTypes.string.isRequired,
 };
