@@ -1,8 +1,8 @@
-import React, { Fragment, useState, useEffect } from "react";
+import React, { useState, useEffect } from "react";
+import PropTypes from 'prop-types';
 import DialogBox from "../DialogBox/DialogBox";
-import Auth from '../../Auth';
 import { config } from "../../settings";
-const auth0=new Auth();
+import { customAlert } from "../Confirm/Confirm";
 
 const EditQuestion = ({
   id,
@@ -10,48 +10,41 @@ const EditQuestion = ({
   activeGameDetails,
   showQuestionEdit,
   onCancel,
-  getQuestions
+  getQuestions,
 }) => {
   const [questionsData, setQuestionsData] = useState(componentData);
-  const {
-    questions,
-    gameId,
-    questionDetail,
-    choices,
-    correctChoice,
-    choicesNameArray,
-    questionId
-  } = questionsData;
+  const convertChoice = value => String.fromCharCode(value + 65);
 
-  const editQuestion = (data = "", id) => {
-    console.log(data);
-    let answers = [];
-    if (data) {
-      data.answers.map((item, index) => {
+  const editQuestion = (questionData = "", questionIdToEdit) => {
+    const answers = [];
+    if (questionData) {
+      questionData.answers.map((item, index) => {
         answers.push({ option: convertChoice(index), value: item });
       });
     }
-    data.answers = answers;
-    console.log("options data -----> ", data);
-    let url = config.baseUrl + "/updatequestion/";
+    const dataWithAnswers = {
+      ...questionData,
+      answers,
+    };
+    const url = `${config.baseUrl}/updatequestion/`;
     fetch(url, {
       method: "POST",
       headers: {
-        authorization: "Bearer " + localStorage.getItem("access_token"),
+        authorization: `Bearer ${localStorage.getItem("access_token")}`,
         Accept: "application/json",
-        "Content-Type": "application/json"
+        "Content-Type": "application/json",
       },
-      body: JSON.stringify({ data: data, id: id })
+      body: JSON.stringify({ data: dataWithAnswers, id: questionIdToEdit }),
     })
       .then(res => res.json())
-      .then(data => {
-        alert(JSON.stringify(data));
+      .then(responseData => {
+        customAlert(JSON.stringify(responseData));
         onCancel();
         setPopupState({ ...popupState, showMessage: false });
 
         getQuestions();
       })
-      .catch(error => console.log(error));
+      .catch(error => console.log(error)); // eslint-disable-line
   };
 
   const [popupState, setPopupState] = useState({
@@ -65,9 +58,18 @@ const EditQuestion = ({
     messageBox: false,
     edit: true,
     create: false,
-    onDelete: null,
-    removeMessage: false
+    onDelete: () => {},
+    removeMessage: false,
   });
+
+  const {
+    questions,
+    questionDetail,
+    correctChoice,
+    choicesNameArray,
+    questionId,
+  } = questionsData;
+
   const {
     showMessage,
     confirmButtonValue,
@@ -80,23 +82,13 @@ const EditQuestion = ({
     edit,
     create,
     onDelete,
-    removeMessage
+    removeMessage,
   } = popupState;
-
-  useEffect(() => {
-    console.log("showquestionedit----------", showQuestionEdit);
-    editHandle(id);
-    // setPopupState({ ...popupState, showMessage: showQuestionEdit });
-  }, []);
-
-  const convertChoice = value => {
-    return String.fromCharCode(value + 65);
-  };
 
   //   Questions and choices
 
   let activeGameName;
-  activeGameDetails.map(item => {
+  activeGameDetails.map(item => { // eslint-disable-line
     if (item.key === "Name") {
       activeGameName = item.value;
     }
@@ -110,13 +102,13 @@ const EditQuestion = ({
         key: "game",
         type: "text",
         title: "Game",
-        value: activeGameName
+        value: activeGameName,
       },
       {
         key: "level",
         type: "text",
         title: "Level",
-        value: questionDetail.difficulty_level
+        value: questionDetail.difficulty_level,
       },
       {
         key: "question",
@@ -124,43 +116,42 @@ const EditQuestion = ({
         title: "Question",
         value: questionDetail.question_statement,
         multiline: true,
-        editable: true
+        editable: true,
       },
       {
         key: "answers",
         type: "options",
         title: "answers",
-        value: choicesNameArray
+        value: choicesNameArray,
       },
       {
         key: "current_choice",
         type: "choice",
         title: "Current choice",
-        value: correctChoice
-      }
-    ]
+        value: correctChoice,
+      },
+    ],
   };
 
-  const getChoices = (questionId, selectedQuestion) => {
-    const url = config.baseUrl + `/choices/${questionId}`;
+  const getChoices = (questionIdToGet, selectedQuestion) => {
+    const url = `${config.baseUrl}/choices/${questionIdToGet}`;
     fetch(url, {
       method: "get",
       headers: {
-        authorization: "Bearer " + localStorage.getItem("access_token"),
+        authorization: `Bearer ${localStorage.getItem("access_token")}`,
         "Content-Type": "Application/json",
-        Accept: "application/json"
-      }
+        Accept: "application/json",
+      },
     })
       .then(res => res.json())
-      .then(data => {
-        console.log("choices data -->", JSON.stringify(data));
-        let choicesName = [];
-        let correctChoice = "";
-        data.map((item, index) => {
+      .then(responseData => {
+        const choicesName = [];
+        let correct = "";
+        responseData.map((item, index) => {
           item.option = convertChoice(index);
           choicesName.push(item.choicestatement);
           if (item.answer === 1) {
-            correctChoice = convertChoice(index);
+            correct = convertChoice(index);
           }
         });
 
@@ -170,27 +161,24 @@ const EditQuestion = ({
           choicesNameArray: choicesName,
           questionDetail: selectedQuestion,
           questionId: selectedQuestion.id,
-          correctChoice: correctChoice
+          correctChoice: correct,
         });
       })
-      .catch(err => console.log(err));
+      .catch(err => console.log(err)); // eslint-disable-line
   };
 
-  //   const onCancel = () => {
-  //     setPopupState({ ...popupState, showMessage: false });
-  //   };
-
-  const editHandle = id => {
-    // console.log("id --- ", id);
-    const selectedQuestion = questions.find(item => {
-      return item.id === id;
-    });
+  const editHandle = questionIdToHandle => {
+    const selectedQuestion = questions.find(item => item.id === questionIdToHandle);
     getChoices(id, selectedQuestion);
-    // console.log("question detail selected", questionDetail);
     setPopupState({ ...popupState, showMessage: showQuestionEdit });
   };
+
+  useEffect(() => {
+    editHandle(id);
+  }, []);
+
   return (
-    <Fragment>
+    <>
       <DialogBox
         confirmButtonValue={confirmButtonValue}
         showMessage={showMessage}
@@ -207,8 +195,26 @@ const EditQuestion = ({
         onDelete={onDelete}
         removeMessage={removeMessage}
       />
-    </Fragment>
+    </>
   );
+};
+
+EditQuestion.propTypes = {
+  id: PropTypes.string,
+  componentData: PropTypes.shape({}),
+  activeGameDetails: PropTypes.arrayOf(PropTypes.shape({})),
+  showQuestionEdit: PropTypes.bool,
+  onCancel: PropTypes.func,
+  getQuestions: PropTypes.func,
+};
+
+EditQuestion.defaultProps = {
+  id: null,
+  componentData: null,
+  activeGameDetails: null,
+  showQuestionEdit: false,
+  onCancel: null,
+  getQuestions: null,
 };
 
 export default EditQuestion;
