@@ -1,30 +1,26 @@
 import React, { useEffect, useState } from "react";
 import ListTable from "../ListTable";
-import Icon from "@material-ui/core/Icon";
-import Auth from "../../Auth";
 import { config } from "../../settings";
 import DialogBox from "../DialogBox/DialogBox";
-import { updatePlayer , deletePlayer } from "./utility";
-
-const auth0 = new Auth();
+import { updatePlayer, deletePlayer } from "./utility";
+import { confirmation } from "../Confirm/Confirm";
 
 const ListPlayers = () => {
-
   const [popupState, setPopupState] = useState({
     showMessage: false,
     confirmButtonValue: "Update",
     messageTitle: "",
     messageDescription: "",
-    onConfirm: "",
+    onConfirm: () => {},
     isConfirmation: true,
     title: "Player detail",
     messageBox: false,
     edit: false,
     create: false,
-    onDelete: null,
-    removeMessage: false
+    onDelete: () => {},
+    removeMessage: false,
   });
-  
+
   const {
     showMessage,
     confirmButtonValue,
@@ -37,16 +33,16 @@ const ListPlayers = () => {
     edit,
     create,
     onDelete,
-    removeMessage
+    removeMessage,
   } = popupState;
-  
+
   const [playerData, setPlayerData] = useState({
     player: [{}],
-    selectedPlayer: { id: "", name: "" }
+    selectedPlayer: { id: "", name: "" },
   });
-  
-  const { player, selectedPlayer } = playerData;
-  
+
+  const { selectedPlayer } = playerData;
+
   const editPlayerFields = {
     id: selectedPlayer.id,
     values: [
@@ -55,86 +51,86 @@ const ListPlayers = () => {
         type: "text",
         title: "Program",
         value: selectedPlayer.program ? selectedPlayer.program : "",
-        editable: true
+        editable: true,
       },
       {
         key: "gender",
         type: "dropdown",
         title: "gender",
-        options:  [{id:"male",title:"male"},{id:"female",title:"female"}],
+        options: [{ id: "male", title: "male" }, { id: "female", title: "female" }],
         editable: true,
-        value: "male"
+        value: "male",
       },
       {
         key: "country",
         type: "text",
         title: "Country",
         value: selectedPlayer.country ? selectedPlayer.country : "",
-        editable: true
+        editable: true,
       },
-      
-    ]
+
+    ],
   };
 
   const columns = [
     {
       name: "Sl. No.",
       selector: "sl",
-      sortable: true
+      sortable: true,
     },
     {
       name: "Id",
       selector: "id",
-      sortable: true
+      sortable: true,
     },
     {
       name: "Last Name",
       selector: "lastname",
-      sortable: true
+      sortable: true,
     },
     {
       name: "First Name",
       selector: "firstname",
-      sortable: true
+      sortable: true,
     },
     {
       name: "Gender",
       selector: "gender",
-      sortable: true
+      sortable: true,
     },
     {
       name: "Country",
       selector: "country",
-      sortable: true
+      sortable: true,
     },
     {
       name: "Program",
       selector: "program",
-      sortable: true
-    }
+      sortable: true,
+    },
   ];
 
   const leadershipcolumns = [
     {
       name: "ID",
       selector: "id",
-      sortable: true
+      sortable: true,
     },
     {
       name: "Name",
       selector: "name",
-      sortable: true
+      sortable: true,
     },
     {
       name: "Rank",
       selector: "rank",
-      sortable: true
+      sortable: true,
     },
     {
       name: "Score",
       selector: "score",
-      sortable: true
-    }
+      sortable: true,
+    },
   ];
 
   const [playersData, setPlayersData] = useState({
@@ -142,76 +138,72 @@ const ListPlayers = () => {
     globalleadership: [{}],
     cohortleadership: [{}],
     noOfPlayers: 0,
-    selectedPlayer: { id: "", name: "" }
+    selectedPlayer: { id: "", name: "" },
   });
   const [cohort, setCohort] = useState(null);
   const [activeTab, setActiveTab] = useState(1);
   // const []
 
-  const { user, globalleadership, cohortleadership, noOfPlayers } = playersData;
+  const {
+    user, globalleadership, cohortleadership, noOfPlayers,
+  } = playersData;
 
   const getPlayers = () => {
-    // console.log(this.props.auth);
-    const url = config.baseUrl + "/users";
+    const url = `${config.baseUrl}/users`;
     fetch(url, {
       method: "get",
       headers: {
-        authorization: "Bearer " + localStorage.getItem("access_token"),
+        authorization: `Bearer ${localStorage.getItem("access_token")}`,
         "Content-Type": "Application/json",
-        Accept: "application/json"
-      }
+        Accept: "application/json",
+      },
     })
       .then(res => res.json())
       .then(data => {
-        let numberOfPlayers = data.length;
-        console.log("api data -->", JSON.stringify(data));
+        const numberOfPlayers = data.length;
         data.map((item, index) => {
           item.sl = index + 1;
         });
         setPlayersData({
           ...playersData,
           user: data,
-          noOfPlayers: numberOfPlayers
+          noOfPlayers: numberOfPlayers,
         });
       })
-      .catch(err => console.log(err));
-    console.log(user);
+      .catch(err => console.log(err)); // eslint-disable-line
   };
   const getCohort = () => {
-    // console.log(this.props.auth);
-    const url = config.baseUrl + "/listCohort";
+    const url = `${config.baseUrl}/listCohort`;
     fetch(url, {
       method: "get",
       headers: {
-        authorization: "Bearer " + localStorage.getItem("access_token"),
+        authorization: `Bearer ${localStorage.getItem("access_token")}`,
         "Content-Type": "Application/json",
-        Accept: "application/json"
-      }
+        Accept: "application/json",
+      },
     })
       .then(res => res.json())
       .then(data => {
-        console.log("api data -->", JSON.stringify(data));
         setCohort(data);
       })
-      .catch(err => console.log(err));
+      .catch(err => console.log(err)); // eslint-disable-line
   };
 
   useEffect(() => {
     getPlayers();
     getCohort();
   }, []);
-  // console.log(cohort, "suyash");
   const handleLeaderShip = (type, id) => {
     let api;
     switch (type) {
       case "all":
-        api = config.baseUrl + "/users";
+        api = `${config.baseUrl}/users`;
         break;
       case "global":
-        api = config.baseUrl + "/list_leaderBoard";
+        api = `${config.baseUrl}/list_leaderBoard`;
         break;
       case "cohort":
-        api = config.baseUrl + `/list_cohort_leaderBoard/${id ? id : "1"}`;
+        api = `${config.baseUrl}/list_cohort_leaderBoard/${id || "1"}`;
         break;
       default:
         break;
@@ -219,19 +211,17 @@ const ListPlayers = () => {
     fetch(api, {
       method: "get",
       headers: {
-        authorization: "Bearer " + localStorage.getItem("access_token"),
+        authorization: `Bearer ${localStorage.getItem("access_token")}`,
         "Content-Type": "Application/json",
-        Accept: "application/json"
-      }
+        Accept: "application/json",
+      },
     })
       .then(res => res.json())
       .then(data => {
-        console.log("api data -->", data);
-
         if (type === "global") {
-          let objArray = [];
+          const objArray = [];
           data.map((item, index) => {
-            let obj = {};
+            const obj = {};
             obj.sl = index + 1;
             obj.id = item.Player.id;
             obj.name = item.Player.firstname;
@@ -242,9 +232,9 @@ const ListPlayers = () => {
           setPlayersData({ ...playersData, globalleadership: objArray });
         }
         if (type === "cohort") {
-          let objArray = [];
+          const objArray = [];
           data.map((item, index) => {
-            let obj = {};
+            const obj = {};
             obj.sl = index + 1;
             obj.id = item.Player.id;
             obj.name = item.Player.firstname;
@@ -261,32 +251,16 @@ const ListPlayers = () => {
           });
           setPlayersData({ ...playersData, user: data });
         }
-
-        console.log("playerrrrrr ---- ", playersData);
       })
-      .catch(err => console.log(err));
+      .catch(err => console.log(err)); // eslint-disable-line
   };
 
   const deleteHandle = playerId => {
-    console.log("cohort id: ", playerId);
-    var r = window.confirm(
-      "Are you sure you want to delete player with id = " + playerId
-    );
-    if (r === true) {
-      deletePlayer(playerId, function() {
-        getPlayers();
-      });
-    } else {
-      return;
-    }
+    confirmation('DemGames', `Are you sure you want to delete player with id = ${playerId}?`, () => deletePlayer(playerId, () => { getPlayers(); }));
   };
 
   const editHandle = id => {
-    // alert("inside edit handle and the player id is -- " + id);
-
-    const selected_player = user.find(item => {
-      return item.id === id;
-    });
+    const selected_player = user.find(item => item.id === id);
 
     setPlayerData({ ...playerData, selectedPlayer: selected_player });
     setPopupState({
@@ -301,17 +275,16 @@ const ListPlayers = () => {
       edit: true,
       create: false,
       onDelete: null,
-      removeMessage: false
+      removeMessage: false,
     });
   };
 
   const onCancel = () => {
     setPopupState({ ...popupState, showMessage: false });
   };
-  
+
   const editPlayer = (data = "", id) => {
-    console.log("dialogbox data", id, data);
-    updatePlayer(data, id, function() {
+    updatePlayer(data, id, () => {
       setPopupState({ ...popupState, showMessage: false });
       getPlayers();
     });
@@ -340,7 +313,8 @@ const ListPlayers = () => {
   // }, []);
 
   return (
-    <><DialogBox
+    <>
+      <DialogBox
         confirmButtonValue={confirmButtonValue}
         showMessage={showMessage}
         messageTitle={messageTitle}
@@ -350,7 +324,6 @@ const ListPlayers = () => {
         onCancel={onCancel}
         title={title}
         data={editPlayerFields}
-     
         messageBox={messageBox}
         edit={edit}
         create={create}
@@ -365,11 +338,13 @@ const ListPlayers = () => {
             <div className="playerbox-value">{noOfPlayers}</div>
           </div>
         </div>
-        <div className="graph"></div>
+        <div className="graph" />
       </div>
       <div className="detail-box">
         <div className="tab-container">
           <div
+            role="button"
+            tabIndex={0}
             className={`tab ${activeTab === 1 ? "active" : ""}`}
             onClick={() => {
               setActiveTab(1);
@@ -379,6 +354,8 @@ const ListPlayers = () => {
             All Players
           </div>
           <div
+            role="button"
+            tabIndex={0}
             className={`tab ${activeTab === 2 ? "active" : ""}`}
             onClick={() => {
               setActiveTab(2);
@@ -388,6 +365,8 @@ const ListPlayers = () => {
             Global Leadership
           </div>
           <div
+            role="button"
+            tabIndex={0}
             className={`tab ${activeTab === 3 ? "active" : ""}`}
             onClick={() => {
               setActiveTab(3);
@@ -422,19 +401,19 @@ const ListPlayers = () => {
                   activeTab === 1
                     ? columns
                     : activeTab === 2
-                    ? leadershipcolumns
-                    : leadershipcolumns,
+                      ? leadershipcolumns
+                      : leadershipcolumns,
                 confirmMsg: "Are you sure you want to delete the player",
                 hasActionBtns: true,
                 data:
                   activeTab === 1
                     ? user
                     : activeTab === 2
-                    ? globalleadership
-                    : cohortleadership,
+                      ? globalleadership
+                      : cohortleadership,
                 callbackAfterDelete: getPlayers,
-                deleteHandle: deleteHandle,
-                editHandle: editHandle
+                deleteHandle,
+                editHandle,
               }}
             />
           </div>
