@@ -1,279 +1,273 @@
 // Get all the choices related to a question
+import { toast } from 'react-toastify';
 import { config } from "../../settings";
-import { ToastContainer, toast } from 'react-toastify';
+
 export const getChoices = async (
   questionId,
   selectedQuestion = null,
-  callbackFunction
+  callbackFunction,
 ) => {
-  const url = config.baseUrl + `/choices/${questionId}`;
+  const url = `${config.baseUrl}/choices/${questionId}`;
   await fetch(url, {
     method: "get",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       "Content-Type": "Application/json",
-      Accept: "application/json"
-    }
+      Accept: "application/json",
+    },
   })
     .then(res => res.json())
     .then(data => {
-      console.log("choices data -->", JSON.stringify(data));
       callbackFunction(data, selectedQuestion);
     })
-    .catch(err => console.log(err));
+    .catch(err => console.log(err)); // eslint-disable-line
 };
 
 // Get the choices referencing(linking) a Quesion
 export const getChoiceLinkingQuestion = (questionId, callbackFunction) => {
-  const url = config.baseUrl + `/choiceLinkingQuestion/${questionId}`;
+  const url = `${config.baseUrl}/choiceLinkingQuestion/${questionId}`;
   fetch(url, {
     method: "get",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       "Content-Type": "Application/json",
-      Accept: "application/json"
-    }
+      Accept: "application/json",
+    },
   })
     .then(res => res.json())
     .then(data => {
-      console.log("choices data -->", JSON.stringify(data));
       callbackFunction(data);
     })
-    .catch(err => console.log(err));
+    .catch(err => console.log(err)); // eslint-disable-line
 };
 
 // Get all the questions related to a game
 export const getQuestions = (gameId, callbackFunction) => {
-  const url = config.baseUrl + `/listquestions/${gameId}`;
+  const url = `${config.baseUrl}/listquestions/${gameId}`;
   fetch(url, {
     method: "get",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       "Content-Type": "Application/json",
-      Accept: "application/json"
-    }
+      Accept: "application/json",
+    },
   })
     .then(res => res.json())
     .then(data => {
-      console.log("api data -->", JSON.stringify(data));
       callbackFunction(data);
     })
-    .catch(err => console.log(err));
+    .catch(err => console.log(err)); // eslint-disable-line
 };
 
 // Add a New Question
 export const addQuestion = (data, callbackFunction) => {
-  let url = config.baseUrl + "/addquestion/";
+  const url = `${config.baseUrl}/addquestion/`;
   fetch(url, {
     method: "POST",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       Accept: "application/json",
-      "Content-Type": "application/json"
+      "Content-Type": "application/json",
     },
-    body: JSON.stringify({ data: data })
+    body: JSON.stringify({ data }),
   })
     .then(res => res.json())
-    .then(data => {
+    .then(() => {
       toast.info("Successfully Added !", {
-        position: toast.POSITION.TOP_CENTER
+        position: toast.POSITION.TOP_CENTER,
       });
       callbackFunction();
     })
-    .catch(err=>toast.error("Sorry...some technical issue !", {
-      position: toast.POSITION.TOP_CENTER
+    .catch(() => toast.error("Sorry...some technical issue !", {
+      position: toast.POSITION.TOP_CENTER,
     }));
 };
 
 // Update Question
 export const updateQuestion = (data = "", id, callbackFunction) => {
-  let url = config.baseUrl + "/updatequestion/";
+  const url = `${config.baseUrl}/updatequestion/`;
   fetch(url, {
     method: "POST",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       Accept: "application/json",
-      "Content-Type": "application/json"
+      "Content-Type": "application/json",
     },
-    body: JSON.stringify({ data: data, id: id })
+    body: JSON.stringify({ data, id }),
   })
     .then(res => res.json())
-    .then(data => {
+    .then(() => {
       toast.info("Updated Successfully !", {
-        position: toast.POSITION.TOP_CENTER
+        position: toast.POSITION.TOP_CENTER,
       });
       callbackFunction();
     })
-    .catch(error => toast.error("Sorry...some technical issue !", {
-      position: toast.POSITION.TOP_CENTER
+    .catch(() => toast.error("Sorry...some technical issue !", {
+      position: toast.POSITION.TOP_CENTER,
     }));
 };
 
 // Delete a question based on ID
 export const deleteQuestion = (questionId, callbackFunction) => {
-  if (window.confirm("Are you sure you want to delete the question")) {
-    let url = config.baseUrl + "/questions/" + questionId;
-    fetch(url, {
-      method: "POST",
-      headers: {
-        authorization: "Bearer " + localStorage.getItem("access_token"),
-        Accept: "application/json",
-        "Content-Type": "application/json"
-      },
-      body: JSON.stringify({ questionId: questionId })
+  const url = `${config.baseUrl}/questions/${questionId}`;
+  fetch(url, {
+    method: "POST",
+    headers: {
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
+      Accept: "application/json",
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({ questionId }),
+  })
+    .then(res => res.json())
+    .then(() => {
+      toast.info("Deleted Successfully !", {
+        position: toast.POSITION.TOP_CENTER,
+      });
+      callbackFunction();
     })
-      .then(res => res.json())
-      .then(data => {
-        toast.info("Deleted Successfully !", {
-          position: toast.POSITION.TOP_CENTER
-        });
-        callbackFunction();
-      })
-      .catch(error => toast.error("Sorry...some technical issue !", {
-        position: toast.POSITION.TOP_CENTER
-      }));
-  }
+    .catch(() => toast.error("Sorry...some technical issue !", {
+      position: toast.POSITION.TOP_CENTER,
+    }));
 };
 
 // Add a New Cohort
 export const addCohort = (data, callbackFunction) => {
-  let url = config.baseUrl + "/AddCohort/";
+  const url = `${config.baseUrl}/AddCohort/`;
   fetch(url, {
     method: "POST",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       Accept: "application/json",
-      "Content-Type": "application/json"
+      "Content-Type": "application/json",
     },
-    body: JSON.stringify({ name: data.name })
+    body: JSON.stringify({ name: data.name }),
   })
     .then(res => {
-      if(res.status == 400){
+      if (res.status == 400) {
         throw new Error();
       }
-      return res.json()
-    }).then(data => {
+      return res.json();
+    }).then(() => {
       toast.info("Successfully Added !", {
-        position: toast.POSITION.TOP_CENTER
+        position: toast.POSITION.TOP_CENTER,
       });
       callbackFunction();
     })
-    .catch(error => toast.error("Sorry...some technical issue !", {
-      position: toast.POSITION.TOP_CENTER
+    .catch(() => toast.error("Sorry...some technical issue !", {
+      position: toast.POSITION.TOP_CENTER,
     }));
 };
 
 // Delete a cohort based on ID
 export const deleteCohort = (cohort_id, callbackFunction) => {
-  let url = config.baseUrl + "/DeleteCohort/";
+  const url = `${config.baseUrl}/DeleteCohort/`;
   fetch(url, {
     method: "POST",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       Accept: "application/json",
-      "Content-Type": "application/json"
+      "Content-Type": "application/json",
     },
-    body: JSON.stringify({ cohort_id: cohort_id })
+    body: JSON.stringify({ cohort_id }),
   })
     .then(res => res.json())
     .then(data => {
-      
-      if(JSON.stringify(data.message) === JSON.stringify("Server Error")) {
+      if (JSON.stringify(data.message) === JSON.stringify("Server Error")) {
         toast.error("Sorry...some technical issue !", {
-          position: toast.POSITION.TOP_CENTER
+          position: toast.POSITION.TOP_CENTER,
         });
       } else {
         toast.info("Deleted Successfully !", {
-          position: toast.POSITION.TOP_CENTER
+          position: toast.POSITION.TOP_CENTER,
         });
-     }
+      }
       callbackFunction();
     })
-    .catch(error => toast.error("Sorry...some technical issue !", {
-      position: toast.POSITION.TOP_CENTER
+    .catch(() => toast.error("Sorry...some technical issue !", {
+      position: toast.POSITION.TOP_CENTER,
     }));
 };
 
 // Update Cohort
 export const updateCohort = (name = "", id, callbackFunction) => {
-  let url = config.baseUrl + "/updatecohort/";
+  const url = `${config.baseUrl}/updatecohort/`;
   fetch(url, {
     method: "POST",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       Accept: "application/json",
-      "Content-Type": "application/json"
+      "Content-Type": "application/json",
     },
-    body: JSON.stringify({ name: name, id: id })
+    body: JSON.stringify({ name, id }),
   })
     .then(res => res.json())
-    .then(data => {
+    .then(() => {
       toast.info("Updated Successfully !", {
-        position: toast.POSITION.TOP_CENTER
+        position: toast.POSITION.TOP_CENTER,
       });
       callbackFunction();
     })
-    .catch(error => toast.error("Sorry...some technical issue !", {
-      position: toast.POSITION.TOP_CENTER
+    .catch(() => toast.error("Sorry...some technical issue !", {
+      position: toast.POSITION.TOP_CENTER,
     }));
 };
 
 // Update Player
 export const updatePlayer = (data, id, callbackFunction) => {
-  let url = config.baseUrl + "/updateUser/";
+  const url = `${config.baseUrl}/updateUser/`;
   fetch(url, {
     method: "POST",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       Accept: "application/json",
-      "Content-Type": "application/json"
+      "Content-Type": "application/json",
     },
-    body: JSON.stringify({ 
-      program: data.program, 
-      id: id,
-      gender : data.gender,
-      country : data.country
-    })
+    body: JSON.stringify({
+      program: data.program,
+      id,
+      gender: data.gender,
+      country: data.country,
+    }),
   })
     .then(res => res.json())
-    .then(data => {
+    .then(() => {
       toast.info("Updated Successfully !", {
-        position: toast.POSITION.TOP_CENTER
+        position: toast.POSITION.TOP_CENTER,
       });
       callbackFunction();
     })
-    .catch(error => toast.error("Sorry...some technical issue !", {
-      position: toast.POSITION.TOP_CENTER
+    .catch(() => toast.error("Sorry...some technical issue !", {
+      position: toast.POSITION.TOP_CENTER,
     }));
 };
 
 // Delete player
 export const deletePlayer = (id, callbackFunction) => {
-  let url = config.baseUrl + "/DeleteUser/";
+  const url = `${config.baseUrl}/DeleteUser/`;
   fetch(url, {
     method: "POST",
     headers: {
-      authorization: "Bearer " + localStorage.getItem("access_token"),
+      authorization: `Bearer ${localStorage.getItem("access_token")}`,
       Accept: "application/json",
-      "Content-Type": "application/json"
+      "Content-Type": "application/json",
     },
-    body: JSON.stringify({ id: id })
+    body: JSON.stringify({ id }),
   })
     .then(res => res.json())
     .then(data => {
-      
-      if(JSON.stringify(data.message) === JSON.stringify("Server Error")) {
+      if (JSON.stringify(data.message) === JSON.stringify("Server Error")) {
         toast.error("Sorry...some technical issue !", {
-          position: toast.POSITION.TOP_CENTER
+          position: toast.POSITION.TOP_CENTER,
         });
       } else {
         toast.info("Deleted Successfully !", {
-          position: toast.POSITION.TOP_CENTER
+          position: toast.POSITION.TOP_CENTER,
         });
-     }
+      }
       callbackFunction();
     })
-    .catch(error => toast.error("Sorry...some technical issue !", {
-      position: toast.POSITION.TOP_CENTER
+    .catch(() => toast.error("Sorry...some technical issue !", {
+      position: toast.POSITION.TOP_CENTER,
     }));
 };
