@@ -1,122 +1,101 @@
-import React, { Fragment, useState, useEffect } from "react";
-import { Link } from "react-router-dom";
-import profileUrl from "../../images/profile.png";
+import React, { useState, useEffect } from "react";
+import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import moment from "moment";
+import profileUrl from "../../images/profile.png";
 import { config } from "../../settings";
 
 import "./styles.scss";
 
 const ProfileInfo = props => {
   let profileProgressData = null;
-  console.log("player email", props.player.player_email);
-  let userEmail = props.player.player_email;
+  const userEmail = props.player.player_email;
 
   const [profileData, setProfileData] = useState({
     progressData: [],
     cohortRank: "0",
     globalRank: "0",
-    cohorts: []
+    cohorts: [],
   });
 
-  const { progressData, cohortRank, globalRank, cohorts } = profileData;
+  const {
+    progressData, cohortRank, globalRank, cohorts,
+  } = profileData;
 
-  const getPlayerProfile = async (email, callbackFunction) => {
-    const url = config.baseUrl + `/user/get_profile/${email}`;
+  const getPlayerProfile = async (email) => {
+    const url = `${config.baseUrl}/user/get_profile/${email}`;
     await fetch(url, {
       method: "get",
       headers: {
-        authorization: "Bearer " + localStorage.getItem("access_token"),
+        authorization: `Bearer ${localStorage.getItem("access_token")}`,
         "Content-Type": "Application/json",
-        Accept: "application/json"
-      }
+        Accept: "application/json",
+      },
     })
       .then(res => res.json())
       .then(data => {
         profileProgressData = data;
-        // console.log("profile data -->", JSON.stringify(data));
-        // callbackFunction(data);
       })
-      .catch(err => console.log(err));
+      .catch(err => console.log(err)); // eslint-disable-line
   };
 
-  //   const setPlayerProgress = data => {
-  //     let filteredData = data.map(item => {
-  //       return {
-  //         gameName: item["Game.caption"],
-  //         score: item.score,
-  //         cohort: item["Cohort.name"],
-  //         cohort_id: item["Cohort.id"],
-  //         playdate: item.playstartdate
-  //       };
-  //     });
-  //     console.log("fil data", filteredData);
-  //     setProfileData({ ...profileData, progressData: filteredData });
-  //   };
-
-  const getCohort = async userEmail => {
-    const url = config.baseUrl + `/listCohort`;
+  const getCohort = async () => {
+    const url = `${config.baseUrl}/listCohort`;
     await fetch(url, {
       method: "get",
       headers: {
-        authorization: "Bearer " + localStorage.getItem("access_token"),
+        authorization: `Bearer ${localStorage.getItem("access_token")}`,
         "Content-Type": "Application/json",
-        Accept: "application/json"
-      }
+        Accept: "application/json",
+      },
     })
       .then(res => res.json())
       .then(data => {
-        console.log("cohortsss data -->", JSON.stringify(data));
         getRank(data[0].id, data);
       })
-      .catch(err => console.log(err));
+      .catch(err => console.log(err)); // eslint-disable-line
   };
 
-  const setRank = (rankObject, cohorts, changeCohort = false) => {
+  const setRank = (rankObject, cohortsToSet, changeCohort = false) => {
     let filteredData = progressData;
     if (!changeCohort) {
-      filteredData = profileProgressData.map(item => {
-        return {
-          gameName: item["Game.caption"],
-          score: item.score,
-          cohort: item["Cohort.name"],
-          cohort_id: item["Cohort.id"],
-          playdate: item.playstartdate
-        };
-      });
+      filteredData = profileProgressData.map(item => ({
+        gameName: item["Game.caption"],
+        score: item.score,
+        cohort: item["Cohort.name"],
+        cohort_id: item["Cohort.id"],
+        playdate: item.playstartdate,
+      }));
     }
     setProfileData({
       ...profileData,
       progressData: filteredData,
       cohortRank: rankObject.cohort_rank,
       globalRank: rankObject.global_rank,
-      cohorts: cohorts
+      cohorts: cohortsToSet,
     });
   };
 
-  const getRank = (cohortId, cohorts, changeCohort = false) => {
-    console.log("selected cohort id: ", cohortId);
-    const url = config.baseUrl + `/get_cohort_rank/${userEmail}/${cohortId}`;
+  const getRank = (cohortId, cohortsToSet, changeCohort = false) => {
+    const url = `${config.baseUrl}/get_cohort_rank/${userEmail}/${cohortId}`;
     fetch(url, {
       method: "get",
       headers: {
-        authorization: "Bearer " + localStorage.getItem("access_token"),
+        authorization: `Bearer ${localStorage.getItem("access_token")}`,
         "Content-Type": "Application/json",
-        Accept: "application/json"
-      }
+        Accept: "application/json",
+      },
     })
       .then(res => res.json())
       .then(data => {
-        console.log("rank data -->", JSON.stringify(data));
-        setRank(data, cohorts, changeCohort);
+        setRank(data, cohortsToSet, changeCohort);
       })
-      .catch(err => console.log(err));
+      .catch(err => console.log(err)); // eslint-disable-line
   };
 
   const handleCohortChange = e => {
-    let cohortId = e.target.value;
+    const cohortId = e.target.value;
     getRank(cohortId, cohorts, true);
-    console.log("coh id: ", cohortId);
   };
 
   useEffect(() => {
@@ -124,10 +103,8 @@ const ProfileInfo = props => {
     getCohort(userEmail);
   }, []);
 
-  console.log("progress data", progressData);
-
   return (
-    <Fragment>
+    <>
       <div className="profile-page">
         <div className="container profile-content">
           <div className="row">
@@ -138,9 +115,11 @@ const ProfileInfo = props => {
                   src={`${props.player.player_picture || profileUrl}`}
                 />
                 <div className="card-body">
-                  <h4 className="card-title text-primary">{`${props.player
-                    .player_given_name || ""} ${props.player
-                    .player_family_name || ""}`}</h4>
+                  <h4 className="card-title text-primary">
+                    {`${props.player
+                      .player_given_name || ""} ${props.player
+                      .player_family_name || ""}`}
+                  </h4>
                   <p className="card-text mb-1 mt-3 text-secondary">
                     {"Email : "}
                     {props.player.player_email || ""}
@@ -149,14 +128,6 @@ const ProfileInfo = props => {
                     {"User Name : "}
                     {props.player.player_username || ""}
                   </p>
-
-                  {/* <Link
-                    to="editprofile"
-                    disabled={true}
-                    className="btn btn-primary w-100 mt-3"
-                  >
-                    Edit Profile
-                  </Link> */}
                 </div>
               </div>
             </div>
@@ -168,13 +139,11 @@ const ProfileInfo = props => {
                     className="custom-select mt-3"
                     onChange={e => handleCohortChange(e)}
                   >
-                    {cohorts.map((item, index) => {
-                      return (
-                        <option key={index} value={item.id}>
-                          {item.name}
-                        </option>
-                      );
-                    })}
+                    {cohorts.map((item) => (
+                      <option key={item.id} value={item.id}>
+                        {item.name}
+                      </option>
+                    ))}
                   </select>
                   <div className="card-body text-center">
                     <h2>Cohort Rank</h2>
@@ -210,21 +179,19 @@ const ProfileInfo = props => {
                           </tr>
                         </thead>
                         <tbody>
-                          {progressData &&
-                            progressData.map((item, index) => {
-                              return (
-                                <tr key={index}>
-                                  <td>{item.gameName}</td>
-                                  <td>{item.score}</td>
-                                  <td>{item.cohort}</td>
-                                  <td>
-                                    {moment(item.playdate).format(
-                                      "Do MMM YYYY, h:mm:ss a"
-                                    )}
-                                  </td>
-                                </tr>
-                              );
-                            })}
+                          {progressData
+                            && progressData.map((item) => (
+                              <tr key={item.playdate}>
+                                <td>{item.gameName}</td>
+                                <td>{item.score}</td>
+                                <td>{item.cohort}</td>
+                                <td>
+                                  {moment(item.playdate).format(
+                                    "Do MMM YYYY, h:mm:ss a",
+                                  )}
+                                </td>
+                              </tr>
+                            ))}
                         </tbody>
                       </table>
                     </div>
@@ -235,16 +202,25 @@ const ProfileInfo = props => {
           </div>
         </div>
       </div>
-    </Fragment>
+    </>
   );
 };
 
 const mapStateToProps = state => ({
-  player: state.authDetail.authDetail
+  player: state.authDetail.authDetail,
 });
 
-// export default connect(mapStateToProps, mapDispatchToProps)(ProfileInfo);
+ProfileInfo.propTypes = {
+  player: PropTypes.shape({
+    player_email: PropTypes.string.isRequired,
+    player_picture: PropTypes.string.isRequired,
+    player_given_name: PropTypes.string.isRequired,
+    player_family_name: PropTypes.string.isRequired,
+    player_username: PropTypes.string.isRequired,
+  }).isRequired,
+};
+
 export default connect(
   mapStateToProps,
-  null
+  null,
 )(ProfileInfo);
