@@ -1,25 +1,19 @@
-import { FETCH_SCORE_DETAIL } from './constants'; 
+import { FETCH_SCORE_DETAIL } from './constants';
 
 const initialState = {
-	scoreDetail: []
+  scoreDetail: [],
 };
 
 const myReducer = (state = initialState, action) => {
-	
-	const newState = { ...state };
-	
-	console.log(newState);
+  const newState = { ...state };
 
-	switch (action.type) {
-	
-		case FETCH_SCORE_DETAIL: state = { ...state, scoreDetail: action.val}
-        return state;
-        
-		default:
-			return newState;
-	}
-	
-	console.log(newState);
+  switch (action.type) {
+    case FETCH_SCORE_DETAIL: state = { ...state, scoreDetail: action.val };
+      return state;
+
+    default:
+      return newState;
+  }
 };
 
 export default myReducer;
