@@ -1,7 +1,17 @@
-import React, { Component, Fragment } from "react";
-import { Route, Switch, Link, BrowserRouter as Router } from "react-router-dom";
+import React, { Component } from "react";
+import PropTypes from 'prop-types';
+import { BrowserRouter as Router } from "react-router-dom";
+import {
+  TabContent,
+  TabPane,
+  Nav,
+  NavItem,
+  NavLink,
+  Row,
+  Col,
+} from "reactstrap";
+import classnames from "classnames";
 import ListPlayers from "../List/ListPlayers";
-import UpdatePlayer from "../../components/Update/UpdateProfile";
 import Register from "../Add/Register";
 import AddGame from "../Add/AddGame";
 import AddChoices from "../Add/AddChoices";
@@ -9,34 +19,10 @@ import AddQuestion from "../Add/AddQuestion";
 import ListQuestions from "../List/ListQuestions";
 import ListChoices from "../List/ListChoices";
 import ListGames from "../List/ListGames";
-import RemovePlayer from "../Remove/RemovePlayer";
-import removequestion from "../Remove/RemoveQuestion";
-import RemoveChoice from "../Remove/RemoveChoice";
-import UpdateGame from "../Update/UpdateGame";
-import UpdateQuestion from "../Update/UpdateQuestion";
-import UpdateChoice from "../Update/UpdateChoice";
-// Bootstrap
-import {
-  TabContent,
-  TabPane,
-  Nav,
-  NavItem,
-  NavLink,
-  Card,
-  Button,
-  CardTitle,
-  CardText,
-  Row,
-  Col
-} from "reactstrap";
-import classnames from "classnames";
-//import NotFound from '../../pages/Landin';
 
 import Auth from "../../Auth";
-//import notfound from './NotFound';
-import Callback from "../../pages/LandingPage/callback";
 
-const auth = new Auth();
+const auth0 = new Auth();
 
 class Admin extends Component {
   constructor(props) {
@@ -56,7 +42,7 @@ class Admin extends Component {
       activeGameTab: "list",
       activePlayerTab: "list",
       activeQuestionTab: "list",
-      activeChoiceTab: "list"
+      activeChoiceTab: "list",
     };
 
     this.handleChange = this.handleChange.bind(this);
@@ -65,19 +51,17 @@ class Admin extends Component {
 
   componentDidMount() {
     if (this.props.email !== null) {
-      const encodedValue = encodeURIComponent(this.state.email);
       fetch(`http://localhost:9000/selectPlayerProfile`, {
         method: "post",
         headers: {
-          authorization: "Bearer "+auth0.getAccessToken(),
+          authorization: `Bearer ${auth0.getAccessToken()}`,
           "Content-Type": "Application/json",
-          Accept: "application/json"
+          Accept: "application/json",
         },
-        body: JSON.stringify(this.state)
+        body: JSON.stringify(this.state),
       })
         .then(res => res.json())
         .then(data => {
-          console.log(data);
           this.setState({
             play_id: data[0].play_id,
             player_id: data[0].player_id,
@@ -91,10 +75,10 @@ class Admin extends Component {
             program: data[0].program,
             program_rank: data[0].program_rank,
             total_rank: data[0].total_rank,
-            email: data[0].email
+            email: data[0].email,
           });
         })
-        .catch(error => console.log(error));
+        .catch(error => console.log(error)); // eslint-disable-line
     }
   }
 
@@ -102,7 +86,7 @@ class Admin extends Component {
     e.preventDefault();
     const sc = e.target.value;
     this.setState({
-      score: sc
+      score: sc,
     });
   }
 
@@ -110,59 +94,61 @@ class Admin extends Component {
     e.preventDefault();
 
     this.setState({
-      total: Number(this.state.score) + Number(this.state.total)
+      total: Number(this.state.score) + Number(this.state.total),
     });
 
     const url = "http://localhost:9000/updateplayerscore";
     fetch(url, {
       method: "POST",
       headers: {
-        authorization: "Bearer "+auth0.getAccessToken(),
+        authorization: `Bearer ${auth0.getAccessToken()}`,
         "Content-Type": "Application/json",
-        Accept: "application/json"
+        Accept: "application/json",
       },
       body: JSON.stringify(this.state),
-      mode: "cors"
+      mode: "cors",
     })
       .then(res => res.json())
-      .then(data => {
-        console.log(data);
-      })
-      .catch(error => console.log(error));
+      .then(data => data)
+      .catch(error => console.log(error)); // eslint-disable-line
   }
 
   toggle(tab) {
     if (this.state.activeTab !== tab) {
       this.setState({
-        activeTab: tab
+        activeTab: tab,
       });
     }
   }
+
   toggleGame(tab) {
     if (this.state.activeGameTab !== tab) {
       this.setState({
-        activeGameTab: tab
+        activeGameTab: tab,
       });
     }
   }
+
   togglePlayer(tab) {
     if (this.state.activePlayerTab !== tab) {
       this.setState({
-        activePlayerTab: tab
+        activePlayerTab: tab,
       });
     }
   }
+
   toggleQuestion(tab) {
     if (this.state.activeQuestionTab !== tab) {
       this.setState({
-        activeQuestionTab: tab
+        activeQuestionTab: tab,
       });
     }
   }
+
   toggleChoice(tab) {
     if (this.state.activeChoiceTab !== tab) {
       this.setState({
-        activeChoiceTab: tab
+        activeChoiceTab: tab,
       });
     }
   }
@@ -170,13 +156,13 @@ class Admin extends Component {
   render() {
     return (
       <Router>
-        <Fragment>
+        <>
           <div className="container">
             <Nav tabs>
               <NavItem>
                 <NavLink
                   className={classnames({
-                    active: this.state.activeTab === "games"
+                    active: this.state.activeTab === "games",
                   })}
                   onClick={() => {
                     this.toggle("games");
@@ -188,7 +174,7 @@ class Admin extends Component {
               <NavItem>
                 <NavLink
                   className={classnames({
-                    active: this.state.activeTab === "players"
+                    active: this.state.activeTab === "players",
                   })}
                   onClick={() => {
                     this.toggle("players");
@@ -200,7 +186,7 @@ class Admin extends Component {
               <NavItem>
                 <NavLink
                   className={classnames({
-                    active: this.state.activeTab === "questions"
+                    active: this.state.activeTab === "questions",
                   })}
                   onClick={() => {
                     this.toggle("questions");
@@ -212,7 +198,7 @@ class Admin extends Component {
               <NavItem>
                 <NavLink
                   className={classnames({
-                    active: this.state.activeTab === "choices"
+                    active: this.state.activeTab === "choices",
                   })}
                   onClick={() => {
                     this.toggle("choices");
@@ -230,27 +216,29 @@ class Admin extends Component {
                       <NavItem>
                         <NavLink
                           className={classnames({
-                            active: this.state.activeGameTab === "list"
+                            active: this.state.activeGameTab === "list",
                           })}
                           onClick={() => {
                             this.toggleGame("list");
                           }}
                         >
                           {" "}
-                          List Games{" "}
+                          List Games
+                          {" "}
                         </NavLink>
                       </NavItem>
                       <NavItem>
                         <NavLink
                           className={classnames({
-                            active: this.state.activeGameTab === "addNew"
+                            active: this.state.activeGameTab === "addNew",
                           })}
                           onClick={() => {
                             this.toggleGame("addNew");
                           }}
                         >
                           {" "}
-                          Add New Game{" "}
+                          Add New Game
+                          {" "}
                         </NavLink>
                       </NavItem>
                     </Nav>
@@ -280,27 +268,29 @@ class Admin extends Component {
                       <NavItem>
                         <NavLink
                           className={classnames({
-                            active: this.state.activePlayerTab === "list"
+                            active: this.state.activePlayerTab === "list",
                           })}
                           onClick={() => {
                             this.togglePlayer("list");
                           }}
                         >
                           {" "}
-                          List Players{" "}
+                          List Players
+                          {" "}
                         </NavLink>
                       </NavItem>
                       <NavItem>
                         <NavLink
                           className={classnames({
-                            active: this.state.activePlayerTab === "addNew"
+                            active: this.state.activePlayerTab === "addNew",
                           })}
                           onClick={() => {
                             this.togglePlayer("addNew");
                           }}
                         >
                           {" "}
-                          Add New Player{" "}
+                          Add New Player
+                          {" "}
                         </NavLink>
                       </NavItem>
                     </Nav>
@@ -330,27 +320,29 @@ class Admin extends Component {
                       <NavItem>
                         <NavLink
                           className={classnames({
-                            active: this.state.activeQuestionTab === "list"
+                            active: this.state.activeQuestionTab === "list",
                           })}
                           onClick={() => {
                             this.toggleQuestion("list");
                           }}
                         >
                           {" "}
-                          List Questions{" "}
+                          List Questions
+                          {" "}
                         </NavLink>
                       </NavItem>
                       <NavItem>
                         <NavLink
                           className={classnames({
-                            active: this.state.activeQuestionTab === "addNew"
+                            active: this.state.activeQuestionTab === "addNew",
                           })}
                           onClick={() => {
                             this.toggleQuestion("addNew");
                           }}
                         >
                           {" "}
-                          Add New Question{" "}
+                          Add New Question
+                          {" "}
                         </NavLink>
                       </NavItem>
                     </Nav>
@@ -380,27 +372,29 @@ class Admin extends Component {
                       <NavItem>
                         <NavLink
                           className={classnames({
-                            active: this.state.activeChoiceTab === "list"
+                            active: this.state.activeChoiceTab === "list",
                           })}
                           onClick={() => {
                             this.toggleChoice("list");
                           }}
                         >
                           {" "}
-                          List Choices{" "}
+                          List Choices
+                          {" "}
                         </NavLink>
                       </NavItem>
                       <NavItem>
                         <NavLink
                           className={classnames({
-                            active: this.state.activeChoiceTab === "addNew"
+                            active: this.state.activeChoiceTab === "addNew",
                           })}
                           onClick={() => {
                             this.toggleChoice("addNew");
                           }}
                         >
                           {" "}
-                          Add New Choice{" "}
+                          Add New Choice
+                          {" "}
                         </NavLink>
                       </NavItem>
                     </Nav>
@@ -424,48 +418,23 @@ class Admin extends Component {
                 </Row>
               </TabPane>
             </TabContent>
-
-            {/* <nav className="navbar navbar-expand-sm bg-dark navbar-dark">
-                            <ul className="navbar-nav">
-                                <li className="nav-item"><Link className="nav-link" to="/list">List players</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/listgames">List games</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/listquestions">List questions</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/listchoices">List choices</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/UpdatePlayer">Update player</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/registerplayer">Register new player</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/addgame">Add game</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/addchoices">Add choices</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/addquestion">Add question</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/removeplayer">Remove player</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/removechoice">Remove choice</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/removequestion">Remove question</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/updategame">Update game</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/updatequestion">Update question</Link> </li>
-                                <li className="nav-item"><Link className="nav-link" to="/updatechoice">Update choice</Link> </li>
-                            </ul>
-                        </nav>
-                        <Switch>
-                            <Route path="/list" component={ListPlayers} />
-                            <Route path="/UpdatePlayer" component={UpdatePlayer} />
-                            <Route path="/registerplayer" component={Register} />
-                            <Route path="/addgame" component={AddGame} />
-                            <Route path="/addchoices" component={AddChoices} />
-                            <Route path="/addquestion" component={AddQuestion} />
-                            <Route path="/listquestions" component={ListQuestions} />
-                            <Route path="/listchoices" component={ListChoices} />
-                            <Route path="/listgames" component={ListGames} />
-                            <Route path="/removeplayer" component={RemovePlayer} />
-                            <Route path="/removechoice" component={RemoveChoice} />
-                            <Route path="/removequestion" component={removequestion} />
-                            <Route path="/updategame" component={UpdateGame} />
-                            <Route path="/updatequestion" component={UpdateQuestion} />
-                            <Route path="/updatechoice" component={UpdateChoice} />
-                        </Switch> */}
           </div>
-        </Fragment>
+        </>
       </Router>
     );
   }
 }
+
+Admin.propTypes = {
+  email: PropTypes.string,
+  given_name: PropTypes.string.isRequired,
+  family_name: PropTypes.string.isRequired,
+  picture: PropTypes.string.isRequired,
+  gender: PropTypes.string.isRequired,
+};
+
+Admin.defaultProps = {
+  email: null,
+};
 
 export default Admin;
