@@ -1,8 +1,4 @@
 import React from "react";
-import ndiLogoUrl from "../../images/ndiLogo.png";
-import arrowBackUrl from "../../images/back.png";
-import profileUrl from "../../images/profile.png";
-import infoUrl from "../../images/info.png";
 import LevelCard from "../../components/LevelCard";
 import "../../commonStyles.scss";
 import "./styles.scss";
@@ -19,7 +15,7 @@ const authDetail = {
   player_email: "",
   player_username: "",
   player_picture: "",
-  player_gender: ""
+  player_gender: "",
 };
 
 class LevelsPage extends React.Component {
@@ -28,23 +24,19 @@ class LevelsPage extends React.Component {
     this.state = { open: false };
   }
 
-  //Handle info icon click to open info dialog box.
+  // Handle info icon click to open info dialog box.
   handleClickOpen = () => {
     this.setState({ open: true });
   };
 
-  componentDidMount() {
-    // console.log(this.props);
-  }
-
-  //Handle info dialog box close.
+  // Handle info dialog box close.
   handleClose = () => {
     this.setState({ open: false });
   };
 
-  //Get list of all modules.
+  // Get list of all modules.
   getModuleNames = () => {
-    const gameData = this.props.gameData.gameData;
+    const { gameData } = this.props.gameData;
     const moduleNames = [];
     gameData.map(modules => {
       moduleNames.push(modules.name);
@@ -54,32 +46,32 @@ class LevelsPage extends React.Component {
 
   // Get Scores for each levels of a particular module.
   getScores = () => {
-    let moduleId = parseInt(this.props.match.params.moduleId);
+    const moduleId = parseInt(this.props.match.params.moduleId, 10);
     const scores = this.props.gameData.scores[moduleId - 1];
     return scores;
   };
 
-  //Get list of par scores for each level of a particular module.
+  // Get list of par scores for each level of a particular module.
   getParScores = () => {
     const { gameData } = this.props;
     let parScores;
-    let moduleId = parseInt(this.props.match.params.moduleId);
+    const moduleId = parseInt(this.props.match.params.moduleId, 10);
     if (gameData[gameData[moduleId - 1]]) {
-      ///gameData[moduleId - 1].levels["0"].par_score = 0;
+      // /gameData[moduleId - 1].levels["0"].par_score = 0;
       parScores = gameData[moduleId - 1].levels.map(level => level.par_score);
-      //parScores[moduleId - 1].levels["0"].par_score = 0;
+      // parScores[moduleId - 1].levels["0"].par_score = 0;
     }
     return parScores;
   };
 
-  //handle Login in action
+  // handle Login in action
   handleLogIn = () => {
     if (!auth0.isAuthenticated()) {
       auth0.login();
     }
   };
 
-  //handle Logout in action
+  // handle Logout in action
   handleLogOut = () => {
     if (auth0.isAuthenticated()) {
       authDetail.player_given_name = "";
@@ -88,8 +80,6 @@ class LevelsPage extends React.Component {
       authDetail.player_username = "";
       authDetail.player_picture = "";
       authDetail.player_gender = "";
-
-      //this.props.clearAuth(authDetail);
       auth0.logout();
     }
   };
@@ -98,16 +88,11 @@ class LevelsPage extends React.Component {
     const moduleNames = this.getModuleNames();
     const { open } = this.state;
     const { gameData } = this.props;
-    console.log("gameDatagameData", gameData);
     const scores = this.getScores();
     const parScores = this.getParScores();
-    const moduleId = parseInt(this.props.match.params.moduleId);
-    console.log(
-      "player_email ---test\n\n",
-      JSON.stringify(this.props.player_email)
-    );
+    const moduleId = parseInt(this.props.match.params.moduleId, 10);
     const moduleName = moduleNames[moduleId - 1];
-    let levels, moduleColor, moduleType;
+    let levels; let moduleColor; let moduleType;
     if (gameData.gameData[moduleId - 1]) {
       levels = gameData.gameData[moduleId - 1].levels;
       moduleColor = this.props.gameData.gameData[moduleId - 1].style;
@@ -117,13 +102,16 @@ class LevelsPage extends React.Component {
     return (
       <div className="landing-page-wrapper">
         <div className="landing-page-container">
-          <p className="game-title"> {moduleName}</p>
+          <p className="game-title">
+            {' '}
+            {moduleName}
+          </p>
           <div className="game-type-card-container">
-            {levels &&
-              levels.length > 0 &&
-              levels.map((data, key) => (
+            {levels
+              && levels.length > 0
+              && levels.map((data) => (
                 <LevelCard
-                  key={key}
+                  key={data.id}
                   level={data.id}
                   moduleId={moduleId}
                   prevLevelScore={data.id > 1 ? scores[data.id - 2] : 0}
@@ -151,31 +139,37 @@ class LevelsPage extends React.Component {
   }
 }
 
-const mapStateToProps = state => {
-  return {
-    player_given_name: state.authDetail.authDetail.player_given_name,
-    player_picture: state.authDetail.authDetail.player_picture,
-    player_email: state.authDetail.authDetail.player_email,
-    gameData: state.gameData
-  };
-};
-
-//Dispatch action to fetch game data and scores.
-const mapDispatchToProps = dispatch => {
-  return {
-    setAuth: authDetail => dispatch(fetchAuthDetails(authDetail)),
-    clearAuth: authDetail => dispatch(clearAuthDetails(authDetail))
-  };
-};
+const mapStateToProps = state => ({
+  player_given_name: state.authDetail.authDetail.player_given_name,
+  player_picture: state.authDetail.authDetail.player_picture,
+  player_email: state.authDetail.authDetail.player_email,
+  gameData: state.gameData,
+});
 
 LevelsPage.propTypes = {
-  gameData: PropTypes.object,
-  match: PropTypes.object
+  gameData: PropTypes.shape({
+    gameData: PropTypes.arrayOf(PropTypes.shape({
+      levels: PropTypes.string,
+      style: PropTypes.string,
+      type: PropTypes.string,
+    })),
+    scores: PropTypes.arrayOf(PropTypes.number),
+    params: PropTypes.shape({
+      moduleId: PropTypes.string,
+    }),
+  }),
+  match: PropTypes.shape({
+    params: PropTypes.shape({
+      moduleId: PropTypes.string,
+    }),
+  }),
+  player_email: PropTypes.string,
 };
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(LevelsPage);
+LevelsPage.defaultProps = {
+  gameData: null,
+  match: null,
+  player_email: null,
+};
 
-// export default LevelsPage;
+export default connect(mapStateToProps)(LevelsPage);
