@@ -1,974 +1,973 @@
 export const gameData = [
-	{
-		id: 1,
-		name: 'Designing an Argument',
-		style: 'blue',
-		type: 'single',
+  {
+    id: 1,
+    name: 'Designing an Argument',
+    style: 'blue',
+    type: 'single',
 
-		levels: [
-			{
-				id: 1,
-				current_score: 0,
-				par_score: 0,
-				total_score: 60,
-				desc: 'Description1',
-				linked_level: 0,
-				questions: [
-					{
-						id: 1,
+    levels: [
+      {
+        id: 1,
+        current_score: 0,
+        par_score: 0,
+        total_score: 60,
+        desc: 'Description1',
+        linked_level: 0,
+        questions: [
+          {
+            id: 1,
 
-						question: 'Resolution',
-						options: [
-							'The topic or claim that is being debated. Example: The death penalty is a justified method of punishment.',
-							'They are laws, decisions of the supreme court and policies of the executive branch that would prevent a new policy from existing.',
-							'This is the Greek word for ethics or character and focuses on demonstrating the reliability, reliability or preparation of the debaters.',
-							'This is the Greek word for logic and focuses on the message and appeals to authority or credibility by offering experience, research or data to support the arguments.'
-						],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 2,
-						question: 'Evidence',
-						options: [
-							'Information used to support a claim.',
-							'They are ways of thinking that would block certain policies or laws. The president (and the cabinet), the leaders of the Congress, the judges of the court, the groups of interest and the public in general have the power to avoid that the policies exist.',
-							'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
-							'A controversial statement that a debater supports or refutes with evidence and reasoning.'
-						],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 3,
-						question: 'Refutation',
-						options: [
-							"The process of attacking an opponent's arguments",
-							'Information used to support a claim.',
-							'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
-							'A controversial statement that a debater supports or refutes with evidence and reasoning.'
-						],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 4,
-						question: 'Claim',
-						options: [
-							'A controversial statement that a debater supports or refutes with evidence and reasoning',
-							"The process of attacking an opponent's arguments",
-							'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
-							'A controversial statement that a debater supports or refutes with evidence and reasoning.'
-						],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 5,
-						question: 'Claim',
-						options: [
-							'A controversial statement that a debater supports or refutes with evidence and reasoning',
-							"The process of attacking an opponent's arguments",
-							'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
-							'A controversial statement that a debater supports or refutes with evidence and reasoning.'
-						],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 6,
-						question: 'Claim',
-						options: [
-							'A controversial statement that a debater supports or refutes with evidence and reasoning',
-							"The process of attacking an opponent's arguments",
-							'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
-							'A controversial statement that a debater supports or refutes with evidence and reasoning.'
-						],
-						correct_answer: [ 0 ]
-					}
-				]
-			},
-			{
-				id: 2,
-				current_score: 0,
-				par_score: 20,
-				total_score: 30,
-				desc: 'Description2',
-				linked_level: 1,
-				questions: [
-					{
-						id: 1,
-						question:
+            question: 'Resolution',
+            options: [
+              'The topic or claim that is being debated. Example: The death penalty is a justified method of punishment.',
+              'They are laws, decisions of the supreme court and policies of the executive branch that would prevent a new policy from existing.',
+              'This is the Greek word for ethics or character and focuses on demonstrating the reliability, reliability or preparation of the debaters.',
+              'This is the Greek word for logic and focuses on the message and appeals to authority or credibility by offering experience, research or data to support the arguments.',
+            ],
+            correct_answer: [0],
+          },
+          {
+            id: 2,
+            question: 'Evidence',
+            options: [
+              'Information used to support a claim.',
+              'They are ways of thinking that would block certain policies or laws. The president (and the cabinet), the leaders of the Congress, the judges of the court, the groups of interest and the public in general have the power to avoid that the policies exist.',
+              'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
+              'A controversial statement that a debater supports or refutes with evidence and reasoning.',
+            ],
+            correct_answer: [0],
+          },
+          {
+            id: 3,
+            question: 'Refutation',
+            options: [
+              "The process of attacking an opponent's arguments",
+              'Information used to support a claim.',
+              'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
+              'A controversial statement that a debater supports or refutes with evidence and reasoning.',
+            ],
+            correct_answer: [0],
+          },
+          {
+            id: 4,
+            question: 'Claim',
+            options: [
+              'A controversial statement that a debater supports or refutes with evidence and reasoning',
+              "The process of attacking an opponent's arguments",
+              'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
+              'A controversial statement that a debater supports or refutes with evidence and reasoning.',
+            ],
+            correct_answer: [0],
+          },
+          {
+            id: 5,
+            question: 'Claim',
+            options: [
+              'A controversial statement that a debater supports or refutes with evidence and reasoning',
+              "The process of attacking an opponent's arguments",
+              'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
+              'A controversial statement that a debater supports or refutes with evidence and reasoning.',
+            ],
+            correct_answer: [0],
+          },
+          {
+            id: 6,
+            question: 'Claim',
+            options: [
+              'A controversial statement that a debater supports or refutes with evidence and reasoning',
+              "The process of attacking an opponent's arguments",
+              'This is the Greek word that means both experience and suffering and during a debate focuses on convincing the audience with emotion.',
+              'A controversial statement that a debater supports or refutes with evidence and reasoning.',
+            ],
+            correct_answer: [0],
+          },
+        ],
+      },
+      {
+        id: 2,
+        current_score: 0,
+        par_score: 20,
+        total_score: 30,
+        desc: 'Description2',
+        linked_level: 1,
+        questions: [
+          {
+            id: 1,
+            question:
 							'The government should implement a tax on pollution to create a healthier environment.',
-						options: [
-							' Ethos (Ethics) - We have an obligation to protect other species and the planet for our children and future generations.',
-							'Pathos (Emotion) - Pollution causes animals to suffocate and die, birds to be covered with tar and unable to fly and disease for our nation.',
-							'Logos (Logic) - Pollution is a systemic risk to our ecosystem, causing thousands of children to become ill each year, which results in millions of dollars in medical bills. '
-						],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 2,
-						question: 'Teachers should be paid a higher wage.',
-						options: [
-							'Ethos (Ethics) - Researchers illustrate that paying teachers more increases long-term socioeconomic gains for students and their families which we need to increase the prosperity of our country.',
-							'Pathos (Emotion) - Teachers often work long hours and spend more time with their students than their own family. Teachers have the added responsibility of educating the future generation.',
-							'Logos (Logic) - 75% of teachers take home 20% less than they need to support themselves and their families.'
-						],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 3,
-						question: 'The death penalty is an unjustified method of punishment.',
-						options: [
-							'Ethos (Ethics) - According to researchers, the death penalty does not deter crime and we as citizens can utilize more human methods to rehabilitate those that break the law.',
-							'Pathos (Emotion) - Innocent lives are taken by the death penalty and families are forced to deal with the trauma of losing a loved one.',
-							'Logos (Logic) - In 10% of death penalty cases, innocent people are mistakenly identified as guilty.'
-						],
-						correct_answer: [ 0 ]
-					}
-				]
-			},
-			{
-				id: 3,
-				current_score: 0,
-				par_score: 30,
-				total_score: 20,
-				desc: 'Description3',
-				linked_level: 2,
-				questions: [
-					{
-						id: 1,
-						question: 'level3 question1',
-						options: [ 'level3  A', 'level3 B', 'level3 C', 'level3 D' ],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 2,
-						question: 'level3 question2',
-						options: [ 'level3 X', 'level3 Y', 'level3 Z', 'level3 A' ],
-						correct_answer: [ 0 ]
-					}
-				]
-			},
-			{
-				id: 4,
-				current_score: 0,
-				par_score: 40,
-				total_score: 50,
-				desc: 'Description4',
-				linked_level: 3,
-				questions: [
-					{
-						id: 1,
-						question: '',
-						options: [ 'level4  A', 'level4 B', 'level4 C', 'level4 D' ],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 2,
-						question: '',
-						options: [ 'level4  A', 'level4 B', 'level4 C', 'level4 D' ],
-						correct_answer: [ 0 ]
-					}
-				]
-			}
-		]
-	},
-	{
-		id: 2,
-		name: 'Finding an evidence',
-		style: 'green',
-		type: 'multi',
+            options: [
+              ' Ethos (Ethics) - We have an obligation to protect other species and the planet for our children and future generations.',
+              'Pathos (Emotion) - Pollution causes animals to suffocate and die, birds to be covered with tar and unable to fly and disease for our nation.',
+              'Logos (Logic) - Pollution is a systemic risk to our ecosystem, causing thousands of children to become ill each year, which results in millions of dollars in medical bills. ',
+            ],
+            correct_answer: [0],
+          },
+          {
+            id: 2,
+            question: 'Teachers should be paid a higher wage.',
+            options: [
+              'Ethos (Ethics) - Researchers illustrate that paying teachers more increases long-term socioeconomic gains for students and their families which we need to increase the prosperity of our country.',
+              'Pathos (Emotion) - Teachers often work long hours and spend more time with their students than their own family. Teachers have the added responsibility of educating the future generation.',
+              'Logos (Logic) - 75% of teachers take home 20% less than they need to support themselves and their families.',
+            ],
+            correct_answer: [0],
+          },
+          {
+            id: 3,
+            question: 'The death penalty is an unjustified method of punishment.',
+            options: [
+              'Ethos (Ethics) - According to researchers, the death penalty does not deter crime and we as citizens can utilize more human methods to rehabilitate those that break the law.',
+              'Pathos (Emotion) - Innocent lives are taken by the death penalty and families are forced to deal with the trauma of losing a loved one.',
+              'Logos (Logic) - In 10% of death penalty cases, innocent people are mistakenly identified as guilty.',
+            ],
+            correct_answer: [0],
+          },
+        ],
+      },
+      {
+        id: 3,
+        current_score: 0,
+        par_score: 30,
+        total_score: 20,
+        desc: 'Description3',
+        linked_level: 2,
+        questions: [
+          {
+            id: 1,
+            question: 'level3 question1',
+            options: ['level3  A', 'level3 B', 'level3 C', 'level3 D'],
+            correct_answer: [0],
+          },
+          {
+            id: 2,
+            question: 'level3 question2',
+            options: ['level3 X', 'level3 Y', 'level3 Z', 'level3 A'],
+            correct_answer: [0],
+          },
+        ],
+      },
+      {
+        id: 4,
+        current_score: 0,
+        par_score: 40,
+        total_score: 50,
+        desc: 'Description4',
+        linked_level: 3,
+        questions: [
+          {
+            id: 1,
+            question: '',
+            options: ['level4  A', 'level4 B', 'level4 C', 'level4 D'],
+            correct_answer: [0],
+          },
+          {
+            id: 2,
+            question: '',
+            options: ['level4  A', 'level4 B', 'level4 C', 'level4 D'],
+            correct_answer: [0],
+          },
+        ],
+      },
+    ],
+  },
+  {
+    id: 2,
+    name: 'Finding an evidence',
+    style: 'green',
+    type: 'multi',
 
-		levels: [
-			{
-				id: 1,
-				current_score: 0,
-				par_score: 0,
-				total_score: 60,
-				desc: 'DescriptionA',
-				linked_level: 0,
-				questions: [
-					{
-						id: 1,
-						question:
+    levels: [
+      {
+        id: 1,
+        current_score: 0,
+        par_score: 0,
+        total_score: 60,
+        desc: 'DescriptionA',
+        linked_level: 0,
+        questions: [
+          {
+            id: 1,
+            question:
 							'What component of the argument (i.e. affirmation, evidence, reasoning) doesn’t match? \n Topic: Death Penalty',
-						options: [
-							'Affirmation: The death penalty should be upheld.',
-							'Evidence: Innocent people are mistakenly killed by the death penalty.',
-							'Reasoning: The criminal justice system makes errors that cannot be remedied once people are killed.'
-						],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 2,
-						question:
+            options: [
+              'Affirmation: The death penalty should be upheld.',
+              'Evidence: Innocent people are mistakenly killed by the death penalty.',
+              'Reasoning: The criminal justice system makes errors that cannot be remedied once people are killed.',
+            ],
+            correct_answer: [0],
+          },
+          {
+            id: 2,
+            question:
 							'What component of the argument (i.e. affirmation, evidence, reasoning) doesn’t match? \n Topic: Environment',
-						options: [
-							'Affirmation:  A certain amount of air pollution is necessary for economic activity.',
-							'Evidence: Technology available today allow firms to manufacture goods and services without releasing harmful toxins into the air.',
-							'Reasoning: In order to have the basic goods and services needed to sustain life, small amounts of pollution is acceptable.'
-						],
-						correct_answer: [ 0 ]
-					},
-					{
-						id: 3,
-						question:
+            options: [
+              'Affirmation:  A certain amount of air pollution is necessary for economic activity.',
+              'Evidence: Technology available today allow firms to manufacture goods and services without releasing harmful toxins into the air.',
+              'Reasoning: In order to have the basic goods and services needed to sustain life, small amounts of pollution is acceptable.',
+            ],
+            correct_answer: [0],
+          },
+          {
+            id: 3,
+            question:
 							'What component of the argument (i.e. affirmation, evidence, reasoning) doesn’t match? \n Topic: Death Teachers',
-						options: [
-							'Affirmation: Teachers need additional support in the classroom.',
-							'Evidence: A higher student teacher ratio results in higher test scores.',
-							'Reasoning: A higher student teacher ratio results in higher test scores.'
-						],
-						correct_answer: [ 0 ]
-					}
-				]
-			},
-			{
-				id: 2,
-				current_score: 0,
-				par_score: 20,
-				total_score: 70,
-				desc: 'DescriptionB',
-				linked_level: 1,
-				questions: [
-					{
-						id: 1,
-						question: 'level2 question1',
-						options: [ 'level2  A', 'level2 B', 'level2 C', 'level2 D' ],
-						correct_answer: [ 0, 1 ]
-					},
-					{
-						id: 2,
-						question: 'level2 question2',
-						options: [ 'level2 X', 'level2 Y', 'level2 Z', 'level2 A' ],
-						correct_answer: [ 0, 1 ]
-					}
-				]
-			},
-			{
-				id: 3,
-				current_score: 0,
-				par_score: 20,
-				total_score: 80,
-				desc: 'DescriptionC',
-				linked_level: 2,
-				questions: [
-					{
-						id: 1,
-						question: 'level3 question1',
-						options: [ 'level3  A', 'level3 B', 'level3 C', 'level3 D' ],
-						correct_answer: [ 0, 1 ]
-					},
-					{
-						id: 2,
-						question: 'level3 question2',
-						options: [ 'level3 X', 'level3 Y', 'level3 Z', 'level3 A' ],
-						correct_answer: [ 0, 1 ]
-					}
-				]
-			}
-		]
-	},
-	{
-		id: 3,
-		name: 'Finding flaws in Argument',
-		style: 'orange',
-		type: 'scenario',
-		levels: [
-			{
-				id: 1,
-				current_score: 0,
-				par_score: 0,
-				total_score: 50,
-				desc: 'DescriptionE',
-				linked_level: 0,
-				questions: [
-					{
-						id: 1,
-						question:
+            options: [
+              'Affirmation: Teachers need additional support in the classroom.',
+              'Evidence: A higher student teacher ratio results in higher test scores.',
+              'Reasoning: A higher student teacher ratio results in higher test scores.',
+            ],
+            correct_answer: [0],
+          },
+        ],
+      },
+      {
+        id: 2,
+        current_score: 0,
+        par_score: 20,
+        total_score: 70,
+        desc: 'DescriptionB',
+        linked_level: 1,
+        questions: [
+          {
+            id: 1,
+            question: 'level2 question1',
+            options: ['level2  A', 'level2 B', 'level2 C', 'level2 D'],
+            correct_answer: [0, 1],
+          },
+          {
+            id: 2,
+            question: 'level2 question2',
+            options: ['level2 X', 'level2 Y', 'level2 Z', 'level2 A'],
+            correct_answer: [0, 1],
+          },
+        ],
+      },
+      {
+        id: 3,
+        current_score: 0,
+        par_score: 20,
+        total_score: 80,
+        desc: 'DescriptionC',
+        linked_level: 2,
+        questions: [
+          {
+            id: 1,
+            question: 'level3 question1',
+            options: ['level3  A', 'level3 B', 'level3 C', 'level3 D'],
+            correct_answer: [0, 1],
+          },
+          {
+            id: 2,
+            question: 'level3 question2',
+            options: ['level3 X', 'level3 Y', 'level3 Z', 'level3 A'],
+            correct_answer: [0, 1],
+          },
+        ],
+      },
+    ],
+  },
+  {
+    id: 3,
+    name: 'Finding flaws in Argument',
+    style: 'orange',
+    type: 'scenario',
+    levels: [
+      {
+        id: 1,
+        current_score: 0,
+        par_score: 0,
+        total_score: 50,
+        desc: 'DescriptionE',
+        linked_level: 0,
+        questions: [
+          {
+            id: 1,
+            question:
 							'Resolution: There should be a youth quota in the national legislative body.You are arguing the affirmative. Your opponents begin their argument by saying the following: Youth quota does not exist because adult decision makers already represent youth concerns. What counterargument would you present?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'In Guatemala, no forum currently exists to bring together decision makers and youth to capture youth concerns.',
-								linked_question: 9
-							},
-							{
-								option:
+                linked_question: 9,
+              },
+              {
+                option:
 									'My parents never accurately represent my concerns. How can adults represent youth concerns?',
-								linked_question: 3
-							}
-						]
-					},
-					{
-						id: 2,
-						question:
+                linked_question: 3,
+              },
+            ],
+          },
+          {
+            id: 2,
+            question:
 							'Your opponents note that in a past assembly meeting, young people were invited to speak about youth issues. This evidence refutes the claim that decision makers only take into account parents’ opinions.The evidence you provided was not strong enough to refute your opponents claims.',
-						options: [ { option: '', linked_question: null }, { option: '', linked_question: null } ],
-						score: 5
-					},
-					{
-						id: 3,
-						question:
+            options: [{ option: '', linked_question: null }, { option: '', linked_question: null }],
+            score: 5,
+          },
+          {
+            id: 3,
+            question:
 							'Your opponents note that in the past 5 assembly meetings, adults have discussed multiple youth issues. Furthermore, one anecdotal example does not refute this evidence and cannot be applied to the population of adult decision-makers as a whole. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'In those past five assembly meetings, four of the five mentions on “youth issues” were voted down without debate by the assembly. ',
-								linked_question: 4
-							},
-							{
-								option:
+                linked_question: 4,
+              },
+              {
+                option:
 									'Parents legally represent young people, how youth issues are addressed in assembly meetings is reflective of parents opinions',
-								linked_question: 2
-							}
-						]
-					},
-					{
-						id: 4,
-						question:
+                linked_question: 2,
+              },
+            ],
+          },
+          {
+            id: 4,
+            question:
 							'Your opponent acknowledges that you are correct about past meetings, however, in the last meeting there was a new amendment passed on youth unemployment. This demonstrates that decision makers are incorporating youth concerns into decision making,',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'In the past five years, amendments passed on youth unemployment have yielded less than a 2% increase in jobs for young people. Passing amendments on youth issues without engaging them meaningfully results in laws that don’t fully address the reality of the problem.',
-								linked_question: 5
-							},
-							{
-								option:
+                linked_question: 5,
+              },
+              {
+                option:
 									'Including youth issues in one amendment does not erase the need for youth quotas. Youth need real representation and a seat at the table.',
-								linked_question: 6
-							}
-						]
-					},
-					{
-						id: 5,
-						question:
+                linked_question: 6,
+              },
+            ],
+          },
+          {
+            id: 5,
+            question:
 							'Your opponent responds by saying change takes time and decision makers in the national assembly have a responsibility to more than just young people who only represent 22% of the population. Positive change for all of society will also mean change for young people.',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'Youth quotas are not only about representing youth issues. Youth quotas represent an opportunity to demonstrate public acknowledgement of equal value and political know how. This contributes to inclusive societal norms and benefits for society as a whole.',
-								linked_question: 7
-							},
-							{
-								option:
+                linked_question: 7,
+              },
+              {
+                option:
 									'Youth candidates will still have expectations and requirements to ensure a quality, capable candidate can be voted for in a democratic process. ',
-								linked_question: 7
-							}
-						]
-					},
-					{
-						id: 6,
-						question:
+                linked_question: 7,
+              },
+            ],
+          },
+          {
+            id: 6,
+            question:
 							'Your opponent responds by saying youth quotas are an ineffective use of assembly seats. Establishing a quota is the opposite of democracy and does not guarantee that a quality, capable candidate will be chosen. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'Youth candidates will still have expectations and requirements to ensure a quality, capable candidate can be voted for in a democratic process. ',
-								linked_question: 7
-							},
-							{
-								option:
+                linked_question: 7,
+              },
+              {
+                option:
 									'Youth quotas are an effective use of assembly seats because they help marginalized populations, such as youth, participate and have their concerns heard.',
-								linked_question: 8
-							}
-						]
-					},
-					{
-						id: 7,
-						question:
+                linked_question: 8,
+              },
+            ],
+          },
+          {
+            id: 7,
+            question:
 							'Your opponent responds by acknowledging the merits of your argument and requesting additional time to conduct further research for a rebuttal.',
-						options: [ { option: '', linked_question: null }, { option: '', linked_question: null } ],
-						score: 10
-					},
-					{
-						id: 8,
-						question:
+            options: [{ option: '', linked_question: null }, { option: '', linked_question: null }],
+            score: 10,
+          },
+          {
+            id: 8,
+            question:
 							'Your opponent responds by saying quotas may give a voice to different populations, but also runs the risk of increased marginalization in the political space. You both agree to conduct more research.',
-						options: [ { option: '', linked_question: null }, { option: '', linked_question: null } ],
-						score: 15
-					},
-					{
-						id: 9,
-						question:
+            options: [{ option: '', linked_question: null }, { option: '', linked_question: null }],
+            score: 15,
+          },
+          {
+            id: 9,
+            question:
 							'Your opponents respond by stating that there are other ways to capture youth concerns--through national surveys for example. How would you respond? ',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'National surveys have limitations and don’t capture information from all youth demographics. Research shows that youth in rural environments are often undersampled and misrepresented.',
-								linked_question: 10
-							},
-							{
-								option:
+                linked_question: 10,
+              },
+              {
+                option:
 									'Simply gathering youth concerns does not ensure decision-maker accountability. They do not capture youth voice.',
-								linked_question: 11
-							}
-						]
-					},
-					{
-						id: 10,
-						question:
+                linked_question: 11,
+              },
+            ],
+          },
+          {
+            id: 10,
+            question:
 							'Your opponents acknowledge the possibility of sampling bias but conclude that youth quotas also do not represent ALL youth concerns. There may be bias among those chosen to serve in the national legislative body.',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'You state that while youth quotas are not able to capture all youth concerns, a small budget could be set aside for a youth and adult coalition to conduct research. This unites members of the national legislative body with a common goal of more inclusively representing constituents and provides space for young elected officials.',
-								linked_question: 14
-							},
-							{
-								option:
+                linked_question: 14,
+              },
+              {
+                option:
 									'National surveys are a good start, but that information does not produce the same outcomes as young people at the table working collaboratively with decisions makers.',
-								linked_question: 13
-							}
-						]
-					},
-					{
-						id: 11,
-						question:
+                linked_question: 13,
+              },
+            ],
+          },
+          {
+            id: 11,
+            question:
 							'Your opponent states that there are other mechanisms in place that ensure accountability and capture the voices of all citizens. The national legislative body recently voted on a measure to increase youth centers in rural communities based on recommendations from community members. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'Occasional outcomes that benefit youth do not demonstrate the full picture. Research states that decision making bodies have miscalculated what youth need repetitively over the past decade resulting in initiatives like a youth center built last year that sits in disuse and disrepair.',
-								linked_question: 13
-							},
-							{
-								option:
+                linked_question: 13,
+              },
+              {
+                option:
 									'Youth centers are not what young people need, they need basic needs like employment and adequate healthcare!',
-								linked_question: 12
-							}
-						]
-					},
-					{
-						id: 12,
-						question:
+                linked_question: 12,
+              },
+            ],
+          },
+          {
+            id: 12,
+            question:
 							'Your opponent states that unemployment and healthcare are priorities for the current administration, as evidenced by the 20% reduction in homelessness and the bill passed for universal healthcare. This demonstrates that youth concerns are being addressed by the national agenda. You believe your opponent’s argument has merit and decide to conduct more research.',
-						options: [
-							{
-								option: '',
-								linked_question: null
-							},
-							{
-								option: '',
-								linked_question: null
-							}
-						],
-						score: 10
-					},
-					{
-						id: 13,
-						question:
+            options: [
+              {
+                option: '',
+                linked_question: null,
+              },
+              {
+                option: '',
+                linked_question: null,
+              },
+            ],
+            score: 10,
+          },
+          {
+            id: 13,
+            question:
 							'Your opponent states that outcomes for youth are measured by outcomes for all of society. For example, in the past year there has been a 20% decrease in homelessness and a bill passed for universal healthcare. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'You state that societal outcomes do not always benefit young people the same as other members of society. Youth quotas can help establish youth-inclusive practices that demonstrate faith in young people’s abilities to contribute.',
-								linked_question: 14
-							},
-							{
-								option:
+                linked_question: 14,
+              },
+              {
+                option:
 									'You state that youth quotas are not only for youth. Youth participation in the national assembly demonstrates the assets and capacity youth have which has positive outcomes on all of society. This is demonstrated by youth quotas in other countries around the world.',
-								linked_question: 15
-							}
-						]
-					},
-					{
-						id: 14,
-						question:
+                linked_question: 15,
+              },
+            ],
+          },
+          {
+            id: 14,
+            question:
 							'Your opponent acknowledges the strength of your argument and the need to do further research to respond.',
-						options: [
-							{
-								option: '',
-								linked_question: null
-							},
-							{
-								option: '',
-								linked_question: null
-							}
-						],
-						score: 15
-					},
-					{
-						id: 15,
-						question:
+            options: [
+              {
+                option: '',
+                linked_question: null,
+              },
+              {
+                option: '',
+                linked_question: null,
+              },
+            ],
+            score: 15,
+          },
+          {
+            id: 15,
+            question:
 							'Your opponent acknowledges the strength of your argument and the need to do further research to respond.',
-						options: [
-							{
-								option: '',
-								linked_question: null
-							},
-							{
-								option: '',
-								linked_question: null
-							}
-						],
-						score: 10
-					}
-				]
-			},
-			{
-				id: 2,
-				current_score: 0,
-				par_score: 40,
-				total_score: 60,
-				desc: 'DescriptionF',
-				linked_level: 0,
-				questions: [
-					{
-						id: 1,
-						question:
+            options: [
+              {
+                option: '',
+                linked_question: null,
+              },
+              {
+                option: '',
+                linked_question: null,
+              },
+            ],
+            score: 10,
+          },
+        ],
+      },
+      {
+        id: 2,
+        current_score: 0,
+        par_score: 40,
+        total_score: 60,
+        desc: 'DescriptionF',
+        linked_level: 0,
+        questions: [
+          {
+            id: 1,
+            question:
 							'Resolution: There should be a youth quota in the national legislative body.You are arguing the affirmative. Your opponents begin their argument by saying the following: Youth quota does not exist because adult decision makers already represent youth concerns. What counterargument would you present?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'In Guatemala, no forum currently exists to bring together decision makers and youth to capture youth concerns.',
-								linked_question: 9
-							},
-							{
-								option:
+                linked_question: 9,
+              },
+              {
+                option:
 									'My parents never accurately represent my concerns. How can adults represent youth concerns?',
-								linked_question: 3
-							}
-						]
-					},
-					{
-						id: 2,
-						question:
+                linked_question: 3,
+              },
+            ],
+          },
+          {
+            id: 2,
+            question:
 							'Your opponents note that in a past assembly meeting, young people were invited to speak about youth issues. This evidence refutes the claim that decision makers only take into account parents’ opinions.The evidence you provided was not strong enough to refute your opponents claims.',
-						options: [ { option: '', linked_question: null }, { option: '', linked_question: null } ],
-						score: 5
-					},
-					{
-						id: 3,
-						question:
+            options: [{ option: '', linked_question: null }, { option: '', linked_question: null }],
+            score: 5,
+          },
+          {
+            id: 3,
+            question:
 							'Your opponents note that in the past 5 assembly meetings, adults have discussed multiple youth issues. Furthermore, one anecdotal example does not refute this evidence and cannot be applied to the population of adult decision-makers as a whole. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'In those past five assembly meetings, four of the five mentions on “youth issues” were voted down without debate by the assembly. ',
-								linked_question: 4
-							},
-							{
-								option:
+                linked_question: 4,
+              },
+              {
+                option:
 									'Parents legally represent young people, how youth issues are addressed in assembly meetings is reflective of parents opinions',
-								linked_question: 2
-							}
-						]
-					},
-					{
-						id: 4,
-						question:
+                linked_question: 2,
+              },
+            ],
+          },
+          {
+            id: 4,
+            question:
 							'Your opponent acknowledges that you are correct about past meetings, however, in the last meeting there was a new amendment passed on youth unemployment. This demonstrates that decision makers are incorporating youth concerns into decision making,',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'In the past five years, amendments passed on youth unemployment have yielded less than a 2% increase in jobs for young people. Passing amendments on youth issues without engaging them meaningfully results in laws that don’t fully address the reality of the problem.',
-								linked_question: 5
-							},
-							{
-								option:
+                linked_question: 5,
+              },
+              {
+                option:
 									'Including youth issues in one amendment does not erase the need for youth quotas. Youth need real representation and a seat at the table.',
-								linked_question: 6
-							}
-						]
-					},
-					{
-						id: 5,
-						question:
+                linked_question: 6,
+              },
+            ],
+          },
+          {
+            id: 5,
+            question:
 							'Your opponent responds by saying change takes time and decision makers in the national assembly have a responsibility to more than just young people who only represent 22% of the population. Positive change for all of society will also mean change for young people.',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'Youth quotas are not only about representing youth issues. Youth quotas represent an opportunity to demonstrate public acknowledgement of equal value and political know how. This contributes to inclusive societal norms and benefits for society as a whole.',
-								linked_question: 7
-							},
-							{
-								option:
+                linked_question: 7,
+              },
+              {
+                option:
 									'Youth candidates will still have expectations and requirements to ensure a quality, capable candidate can be voted for in a democratic process. ',
-								linked_question: 7
-							}
-						]
-					},
-					{
-						id: 6,
-						question:
+                linked_question: 7,
+              },
+            ],
+          },
+          {
+            id: 6,
+            question:
 							'Your opponent responds by saying youth quotas are an ineffective use of assembly seats. Establishing a quota is the opposite of democracy and does not guarantee that a quality, capable candidate will be chosen. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'Youth candidates will still have expectations and requirements to ensure a quality, capable candidate can be voted for in a democratic process. ',
-								linked_question: 7
-							},
-							{
-								option:
+                linked_question: 7,
+              },
+              {
+                option:
 									'Youth quotas are an effective use of assembly seats because they help marginalized populations, such as youth, participate and have their concerns heard.',
-								linked_question: 8
-							}
-						]
-					},
-					{
-						id: 7,
-						question:
+                linked_question: 8,
+              },
+            ],
+          },
+          {
+            id: 7,
+            question:
 							'Your opponent responds by acknowledging the merits of your argument and requesting additional time to conduct further research for a rebuttal.',
-						options: [ { option: '', linked_question: null }, { option: '', linked_question: null } ],
-						score: 10
-					},
-					{
-						id: 8,
-						question:
+            options: [{ option: '', linked_question: null }, { option: '', linked_question: null }],
+            score: 10,
+          },
+          {
+            id: 8,
+            question:
 							'Your opponent responds by saying quotas may give a voice to different populations, but also runs the risk of increased marginalization in the political space. You both agree to conduct more research.',
-						options: [ { option: '', linked_question: null }, { option: '', linked_question: null } ],
-						score: 15
-					},
-					{
-						id: 9,
-						question:
+            options: [{ option: '', linked_question: null }, { option: '', linked_question: null }],
+            score: 15,
+          },
+          {
+            id: 9,
+            question:
 							'Your opponents respond by stating that there are other ways to capture youth concerns--through national surveys for example. How would you respond? ',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'National surveys have limitations and don’t capture information from all youth demographics. Research shows that youth in rural environments are often undersampled and misrepresented.',
-								linked_question: 10
-							},
-							{
-								option:
+                linked_question: 10,
+              },
+              {
+                option:
 									'Simply gathering youth concerns does not ensure decision-maker accountability. They do not capture youth voice.',
-								linked_question: 11
-							}
-						]
-					},
-					{
-						id: 10,
-						question:
+                linked_question: 11,
+              },
+            ],
+          },
+          {
+            id: 10,
+            question:
 							'Your opponents acknowledge the possibility of sampling bias but conclude that youth quotas also do not represent ALL youth concerns. There may be bias among those chosen to serve in the national legislative body.',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'You state that while youth quotas are not able to capture all youth concerns, a small budget could be set aside for a youth and adult coalition to conduct research. This unites members of the national legislative body with a common goal of more inclusively representing constituents and provides space for young elected officials.',
-								linked_question: 14
-							},
-							{
-								option:
+                linked_question: 14,
+              },
+              {
+                option:
 									'National surveys are a good start, but that information does not produce the same outcomes as young people at the table working collaboratively with decisions makers.',
-								linked_question: 13
-							}
-						]
-					},
-					{
-						id: 11,
-						question:
+                linked_question: 13,
+              },
+            ],
+          },
+          {
+            id: 11,
+            question:
 							'Your opponent states that there are other mechanisms in place that ensure accountability and capture the voices of all citizens. The national legislative body recently voted on a measure to increase youth centers in rural communities based on recommendations from community members. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'Occasional outcomes that benefit youth do not demonstrate the full picture. Research states that decision making bodies have miscalculated what youth need repetitively over the past decade resulting in initiatives like a youth center built last year that sits in disuse and disrepair.',
-								linked_question: 13
-							},
-							{
-								option:
+                linked_question: 13,
+              },
+              {
+                option:
 									'Youth centers are not what young people need, they need basic needs like employment and adequate healthcare!',
-								linked_question: 12
-							}
-						]
-					},
-					{
-						id: 12,
-						question:
+                linked_question: 12,
+              },
+            ],
+          },
+          {
+            id: 12,
+            question:
 							'Your opponent states that unemployment and healthcare are priorities for the current administration, as evidenced by the 20% reduction in homelessness and the bill passed for universal healthcare. This demonstrates that youth concerns are being addressed by the national agenda. You believe your opponent’s argument has merit and decide to conduct more research.',
-						options: [
-							{
-								option: '',
-								linked_question: null
-							},
-							{
-								option: '',
-								linked_question: null
-							}
-						],
-						score: 10
-					},
-					{
-						id: 13,
-						question:
+            options: [
+              {
+                option: '',
+                linked_question: null,
+              },
+              {
+                option: '',
+                linked_question: null,
+              },
+            ],
+            score: 10,
+          },
+          {
+            id: 13,
+            question:
 							'Your opponent states that outcomes for youth are measured by outcomes for all of society. For example, in the past year there has been a 20% decrease in homelessness and a bill passed for universal healthcare. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'You state that societal outcomes do not always benefit young people the same as other members of society. Youth quotas can help establish youth-inclusive practices that demonstrate faith in young people’s abilities to contribute.',
-								linked_question: 14
-							},
-							{
-								option:
+                linked_question: 14,
+              },
+              {
+                option:
 									'You state that youth quotas are not only for youth. Youth participation in the national assembly demonstrates the assets and capacity youth have which has positive outcomes on all of society. This is demonstrated by youth quotas in other countries around the world.',
-								linked_question: 15
-							}
-						]
-					},
-					{
-						id: 14,
-						question:
+                linked_question: 15,
+              },
+            ],
+          },
+          {
+            id: 14,
+            question:
 							'Your opponent acknowledges the strength of your argument and the need to do further research to respond.',
-						options: [
-							{
-								option: '',
-								linked_question: null
-							},
-							{
-								option: '',
-								linked_question: null
-							}
-						],
-						score: 15
-					},
-					{
-						id: 15,
-						question:
+            options: [
+              {
+                option: '',
+                linked_question: null,
+              },
+              {
+                option: '',
+                linked_question: null,
+              },
+            ],
+            score: 15,
+          },
+          {
+            id: 15,
+            question:
 							'Your opponent acknowledges the strength of your argument and the need to do further research to respond.',
-						options: [
-							{
-								option: '',
-								linked_question: null
-							},
-							{
-								option: '',
-								linked_question: null
-							}
-						],
-						score: 10
-					}
-				]
-			},
-			{
-				id: 3,
-				current_score: 0,
-				par_score: 10,
-				total_score: 70,
-				desc: 'DescriptionF',
-				linked_level: 0,
-				questions: [
-					{
-						id: 1,
-						question:
+            options: [
+              {
+                option: '',
+                linked_question: null,
+              },
+              {
+                option: '',
+                linked_question: null,
+              },
+            ],
+            score: 10,
+          },
+        ],
+      },
+      {
+        id: 3,
+        current_score: 0,
+        par_score: 10,
+        total_score: 70,
+        desc: 'DescriptionF',
+        linked_level: 0,
+        questions: [
+          {
+            id: 1,
+            question:
 							'Resolution: There should be a youth quota in the national legislative body.You are arguing the affirmative. Your opponents begin their argument by saying the following: Youth quota does not exist because adult decision makers already represent youth concerns. What counterargument would you present?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'In Guatemala, no forum currently exists to bring together decision makers and youth to capture youth concerns.',
-								linked_question: 9
-							},
-							{
-								option:
+                linked_question: 9,
+              },
+              {
+                option:
 									'My parents never accurately represent my concerns. How can adults represent youth concerns?',
-								linked_question: 3
-							}
-						]
-					},
-					{
-						id: 2,
-						question:
+                linked_question: 3,
+              },
+            ],
+          },
+          {
+            id: 2,
+            question:
 							'Your opponents note that in a past assembly meeting, young people were invited to speak about youth issues. This evidence refutes the claim that decision makers only take into account parents’ opinions.The evidence you provided was not strong enough to refute your opponents claims.',
-						options: [ { option: '', linked_question: null }, { option: '', linked_question: null } ],
-						score: 5
-					},
-					{
-						id: 3,
-						question:
+            options: [{ option: '', linked_question: null }, { option: '', linked_question: null }],
+            score: 5,
+          },
+          {
+            id: 3,
+            question:
 							'Your opponents note that in the past 5 assembly meetings, adults have discussed multiple youth issues. Furthermore, one anecdotal example does not refute this evidence and cannot be applied to the population of adult decision-makers as a whole. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'In those past five assembly meetings, four of the five mentions on “youth issues” were voted down without debate by the assembly. ',
-								linked_question: 4
-							},
-							{
-								option:
+                linked_question: 4,
+              },
+              {
+                option:
 									'Parents legally represent young people, how youth issues are addressed in assembly meetings is reflective of parents opinions',
-								linked_question: 2
-							}
-						]
-					},
-					{
-						id: 4,
-						question:
+                linked_question: 2,
+              },
+            ],
+          },
+          {
+            id: 4,
+            question:
 							'Your opponent acknowledges that you are correct about past meetings, however, in the last meeting there was a new amendment passed on youth unemployment. This demonstrates that decision makers are incorporating youth concerns into decision making,',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'In the past five years, amendments passed on youth unemployment have yielded less than a 2% increase in jobs for young people. Passing amendments on youth issues without engaging them meaningfully results in laws that don’t fully address the reality of the problem.',
-								linked_question: 5
-							},
-							{
-								option:
+                linked_question: 5,
+              },
+              {
+                option:
 									'Including youth issues in one amendment does not erase the need for youth quotas. Youth need real representation and a seat at the table.',
-								linked_question: 6
-							}
-						]
-					},
-					{
-						id: 5,
-						question:
+                linked_question: 6,
+              },
+            ],
+          },
+          {
+            id: 5,
+            question:
 							'Your opponent responds by saying change takes time and decision makers in the national assembly have a responsibility to more than just young people who only represent 22% of the population. Positive change for all of society will also mean change for young people.',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'Youth quotas are not only about representing youth issues. Youth quotas represent an opportunity to demonstrate public acknowledgement of equal value and political know how. This contributes to inclusive societal norms and benefits for society as a whole.',
-								linked_question: 7
-							},
-							{
-								option:
+                linked_question: 7,
+              },
+              {
+                option:
 									'Youth candidates will still have expectations and requirements to ensure a quality, capable candidate can be voted for in a democratic process. ',
-								linked_question: 7
-							}
-						]
-					},
-					{
-						id: 6,
-						question:
+                linked_question: 7,
+              },
+            ],
+          },
+          {
+            id: 6,
+            question:
 							'Your opponent responds by saying youth quotas are an ineffective use of assembly seats. Establishing a quota is the opposite of democracy and does not guarantee that a quality, capable candidate will be chosen. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'Youth candidates will still have expectations and requirements to ensure a quality, capable candidate can be voted for in a democratic process. ',
-								linked_question: 7
-							},
-							{
-								option:
+                linked_question: 7,
+              },
+              {
+                option:
 									'Youth quotas are an effective use of assembly seats because they help marginalized populations, such as youth, participate and have their concerns heard.',
-								linked_question: 8
-							}
-						]
-					},
-					{
-						id: 7,
-						question:
+                linked_question: 8,
+              },
+            ],
+          },
+          {
+            id: 7,
+            question:
 							'Your opponent responds by acknowledging the merits of your argument and requesting additional time to conduct further research for a rebuttal.',
-						options: [ { option: '', linked_question: null }, { option: '', linked_question: null } ],
-						score: 10
-					},
-					{
-						id: 8,
-						question:
+            options: [{ option: '', linked_question: null }, { option: '', linked_question: null }],
+            score: 10,
+          },
+          {
+            id: 8,
+            question:
 							'Your opponent responds by saying quotas may give a voice to different populations, but also runs the risk of increased marginalization in the political space. You both agree to conduct more research.',
-						options: [ { option: '', linked_question: null }, { option: '', linked_question: null } ],
-						score: 15
-					},
-					{
-						id: 9,
-						question:
+            options: [{ option: '', linked_question: null }, { option: '', linked_question: null }],
+            score: 15,
+          },
+          {
+            id: 9,
+            question:
 							'Your opponents respond by stating that there are other ways to capture youth concerns--through national surveys for example. How would you respond? ',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'National surveys have limitations and don’t capture information from all youth demographics. Research shows that youth in rural environments are often undersampled and misrepresented.',
-								linked_question: 10
-							},
-							{
-								option:
+                linked_question: 10,
+              },
+              {
+                option:
 									'Simply gathering youth concerns does not ensure decision-maker accountability. They do not capture youth voice.',
-								linked_question: 11
-							}
-						]
-					},
-					{
-						id: 10,
-						question:
+                linked_question: 11,
+              },
+            ],
+          },
+          {
+            id: 10,
+            question:
 							'Your opponents acknowledge the possibility of sampling bias but conclude that youth quotas also do not represent ALL youth concerns. There may be bias among those chosen to serve in the national legislative body.',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'You state that while youth quotas are not able to capture all youth concerns, a small budget could be set aside for a youth and adult coalition to conduct research. This unites members of the national legislative body with a common goal of more inclusively representing constituents and provides space for young elected officials.',
-								linked_question: 14
-							},
-							{
-								option:
+                linked_question: 14,
+              },
+              {
+                option:
 									'National surveys are a good start, but that information does not produce the same outcomes as young people at the table working collaboratively with decisions makers.',
-								linked_question: 13
-							}
-						]
-					},
-					{
-						id: 11,
-						question:
+                linked_question: 13,
+              },
+            ],
+          },
+          {
+            id: 11,
+            question:
 							'Your opponent states that there are other mechanisms in place that ensure accountability and capture the voices of all citizens. The national legislative body recently voted on a measure to increase youth centers in rural communities based on recommendations from community members. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'Occasional outcomes that benefit youth do not demonstrate the full picture. Research states that decision making bodies have miscalculated what youth need repetitively over the past decade resulting in initiatives like a youth center built last year that sits in disuse and disrepair.',
-								linked_question: 13
-							},
-							{
-								option:
+                linked_question: 13,
+              },
+              {
+                option:
 									'Youth centers are not what young people need, they need basic needs like employment and adequate healthcare!',
-								linked_question: 12
-							}
-						]
-					},
-					{
-						id: 12,
-						question:
+                linked_question: 12,
+              },
+            ],
+          },
+          {
+            id: 12,
+            question:
 							'Your opponent states that unemployment and healthcare are priorities for the current administration, as evidenced by the 20% reduction in homelessness and the bill passed for universal healthcare. This demonstrates that youth concerns are being addressed by the national agenda. You believe your opponent’s argument has merit and decide to conduct more research.',
-						options: [
-							{
-								option: '',
-								linked_question: null
-							},
-							{
-								option: '',
-								linked_question: null
-							}
-						],
-						score: 10
-					},
-					{
-						id: 13,
-						question:
+            options: [
+              {
+                option: '',
+                linked_question: null,
+              },
+              {
+                option: '',
+                linked_question: null,
+              },
+            ],
+            score: 10,
+          },
+          {
+            id: 13,
+            question:
 							'Your opponent states that outcomes for youth are measured by outcomes for all of society. For example, in the past year there has been a 20% decrease in homelessness and a bill passed for universal healthcare. How do you respond?',
-						options: [
-							{
-								option:
+            options: [
+              {
+                option:
 									'You state that societal outcomes do not always benefit young people the same as other members of society. Youth quotas can help establish youth-inclusive practices that demonstrate faith in young people’s abilities to contribute.',
-								linked_question: 14
-							},
-							{
-								option:
+                linked_question: 14,
+              },
+              {
+                option:
 									'You state that youth quotas are not only for youth. Youth participation in the national assembly demonstrates the assets and capacity youth have which has positive outcomes on all of society. This is demonstrated by youth quotas in other countries around the world.',
-								linked_question: 15
-							}
-						]
-					},
-					{
-						id: 14,
-						question:
+                linked_question: 15,
+              },
+            ],
+          },
+          {
+            id: 14,
+            question:
 							'Your opponent acknowledges the strength of your argument and the need to do further research to respond.',
-						options: [
-							{
-								option: '',
-								linked_question: null
-							},
-							{
-								option: '',
-								linked_question: null
-							}
-						],
-						score: 15
-					},
-					{
-						id: 15,
-						question:
+            options: [
+              {
+                option: '',
+                linked_question: null,
+              },
+              {
+                option: '',
+                linked_question: null,
+              },
+            ],
+            score: 15,
+          },
+          {
+            id: 15,
+            question:
 							'Your opponent acknowledges the strength of your argument and the need to do further research to respond.',
-						options: [
-							{
-								option: '',
-								linked_question: null
-							},
-							{
-								option: '',
-								linked_question: null
-							}
-						],
-						score: 10
-					}
-				]
-			}
-		]
-	}
+            options: [
+              {
+                option: '',
+                linked_question: null,
+              },
+              {
+                option: '',
+                linked_question: null,
+              },
+            ],
+            score: 10,
+          },
+        ],
+      },
+    ],
+  },
 ];
-	
