@@ -1,48 +1,35 @@
 import React from "react";
-import ndiLogoUrl from "../../images/ndiLogo.png";
-import arrowBackUrl from "../../images/back.png";
-import profileUrl from "../../images/profile.png";
-import infoUrl from "../../images/info.png";
+import PropTypes from 'prop-types';
 import LevelCard from "../../components/LevelCard";
 import "../../commonStyles.scss";
 import "../LevelsPage/styles.scss";
-import { connect } from "react-redux";
-import GameInfo from "../../components/GameInfo";
-import PropTypes, { element } from "prop-types";
 import Auth from "../../Auth";
 import { config } from "../../settings";
 
 const auth0 = new Auth();
 
-const authDetail = {
-  player_given_name: "",
-  player_family_name: "",
-  player_email: "",
-  player_username: "",
-  player_picture: "",
-  player_gender: ""
-};
 class ScenarioLevels extends React.Component {
   state = { levels: 0, game: {} };
+
   componentDidMount() {
-    const url = config.baseUrl + `/level_by_game/${this.props.match.params.moduleId}`;
+    const url = `${config.baseUrl}/level_by_game/${this.props.match.params.moduleId}`;
     fetch(url, {
       method: "GET",
       headers: {
-        authorization: "Bearer " + auth0.getAccessToken(),
+        authorization: `Bearer ${auth0.getAccessToken()}`,
         Accept: "application/json",
-        "Content-Type": "application/json"
-      }
+        "Content-Type": "application/json",
+      },
     })
       .then(res => res.json())
       .then(data => {
         this.setState({ levels: data.length, game: data.game });
-        console.log("level data", data);
       })
-      .catch(error => console.log(error));
+      .catch(error => console.log(error)); // eslint-disable-line
   }
+
   getLevelComponents() {
-    let elements = [];
+    const elements = [];
     for (let x = 0; x < this.state.levels; x++) {
       elements.push(
         <LevelCard
@@ -51,38 +38,22 @@ class ScenarioLevels extends React.Component {
           moduleId={this.props.match.params.moduleId}
           showScore={false}
           moduleType="scenario"
-          moduleColor={this.state.game.style === null?"orange": this.state.game.style}
-        />
+          moduleColor={this.state.game.style === null ? "orange" : this.state.game.style}
+        />,
       );
     }
     return elements;
   }
+
   render() {
     return (
       <div className="landing-page-wrapper">
         <div className="landing-page-container">
-          <p className="game-title"> {this.state.game.caption}</p>
+          <p className="game-title">
+            {' '}
+            {this.state.game.caption}
+          </p>
           <div className="game-type-card-container">
-            {/* {levels &&
-		    				levels.length > 0 &&
-		    				levels.map((data, key) => (
-		    					<LevelCard
-		    						key={key}
-		    						level={data.id}
-		    						moduleId={moduleId}
-		    						prevLevelScore={data.id > 1 ? scores[data.id - 2] : 0}
-		    						currentScore={scores[data.id - 1]}
-		    						parScore={parScores ? parScores[data.id - 1]: null}
-		    						linkedLevel={data.linked_level}
-		    						description={data.desc}
-		    						totalScore={data.total_score}
-		    						questions={data.questions}
-		    						moduleName={moduleName}
-		    						moduleType={moduleType}
-		    						moduleColor={moduleColor}
-		    						player_email={this.props.player_email == null ? "default":this.props.player_email }
-		    					/>
-                            ))} */}
             {this.getLevelComponents()}
           </div>
         </div>
@@ -90,5 +61,13 @@ class ScenarioLevels extends React.Component {
     );
   }
 }
+
+ScenarioLevels.propTypes = {
+  match: PropTypes.shape({
+    params: PropTypes.shape({
+      moduleId: PropTypes.string,
+    }),
+  }).isRequired,
+};
 
 export default ScenarioLevels;
