@@ -3,6 +3,6 @@ import alert from './alert';
 import myReducer from '../pages/LandingPage/reducers';
 
 export default combineReducers({
-    alert , 
-    myReducer
+  alert,
+  myReducer,
 });
