@@ -1,13 +1,13 @@
 let config = {};
 
 if (process.env.NODE_ENV === 'development') {
-	config = { baseUrl: 'http://localhost:9000/api' };
-	// 192.168.13.86 is ip of network you are connected to change it with your ip address.
-	// config = { baseUrl: 'http://192.168.13.86:9000' };
+  config = { baseUrl: 'http://localhost:9000/api' };
+  // 192.168.13.86 is ip of network you are connected to change it with your ip address.
+  // config = { baseUrl: 'http://192.168.13.86:9000' };
 } else if (process.env.NODE_ENV === 'production') {
-	config = { baseUrl: '' };
+  config = { baseUrl: '' };
 } else {
-	config = { baseUrl: 'http://localhost:9000/api' };
+  config = { baseUrl: 'http://localhost:9000/api' };
 }
 
 // provide redirect uri
