@@ -2,33 +2,31 @@ import { createStore } from 'redux';
 import rootReducer from './reducers';
 
 function saveToLocalStorage(state) {
-	try {
-		const serializedState = JSON.stringify(state);
-		localStorage.setItem('state', serializedState);
-	} catch (err) {
-		console.log(err);
-	}
+  try {
+    const serializedState = JSON.stringify(state);
+    localStorage.setItem('state', serializedState);
+  } catch (err) {
+		console.log(err); // eslint-disable-line
+  }
 }
 
 function loadFromLocalStorage() {
-	try {
-		const serializedState = localStorage.getItem('state');
-		if (serializedState === null) return undefined;
-		return JSON.parse(serializedState);
-	} catch (err) {
-		console.log(err);
-		return undefined;
-	}
+  try {
+    const serializedState = localStorage.getItem('state');
+    if (serializedState === null) return undefined;
+    return JSON.parse(serializedState);
+  } catch (err) {
+		console.log(err); // eslint-disable-line
+    return undefined;
+  }
 }
-
 
 export const persistedState = loadFromLocalStorage();
 
-
 const store = createStore(
-	rootReducer,
-	persistedState,
-	window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+  rootReducer,
+  persistedState,
+  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
 );
 
 store.subscribe(() => saveToLocalStorage(store.getState()));
