@@ -4,41 +4,33 @@ import thunk from 'redux-thunk';
 import rootReducer from './reducers';
 
 function saveToLocalStorage(state) {
-	try {
-		const serializedState = JSON.stringify(state);
-		localStorage.setItem('state', serializedState);
-	} catch (err) {
-		console.log(err);
-	}
+  try {
+    const serializedState = JSON.stringify(state);
+    localStorage.setItem('state', serializedState);
+  } catch (err) {
+		console.log(err); // eslint-disable-line
+  }
 }
 
 function loadFromLocalStorage() {
-	try {
-		const serializedState = localStorage.getItem('state');
-		if (serializedState === null) return undefined;
-		return JSON.parse(serializedState);
-	} catch (err) {
-		console.log(err);
-		return undefined;
-	}
+  try {
+    const serializedState = localStorage.getItem('state');
+    if (serializedState === null) return undefined;
+    return JSON.parse(serializedState);
+  } catch (err) {
+		console.log(err);  // eslint-disable-line
+    return undefined;
+  }
 }
 
-
 export const persistedState = loadFromLocalStorage();
-
-
-
-
-
-
-const initialState = {};
 
 const middleware = [thunk];
 
 const store = createStore(
-    rootReducer,
-    persistedState,
-    composeWithDevTools(applyMiddleware(...middleware))
+  rootReducer,
+  persistedState,
+  composeWithDevTools(applyMiddleware(...middleware)),
 );
 store.subscribe(() => saveToLocalStorage(store.getState()));
 
